Move code out of BoogieGenerator enough that it gets below 5k lines and Rider starts showing normal hints in it again
