+++ Source/DafnyCore/Verifier/BoogieGenerator.cs
@@ -1929,60 +1929,6 @@ namespace Microsoft.Dafny {
       return assumeCmd;
     }
 
-    private void AddFunctionOverrideEnsChk(Function f, BoogieStmtListBuilder builder, ExpressionTranslator etran,
-      Dictionary<IVariable, Expression> substMap,
-      Dictionary<TypeParameter, Type> typeMap,
-      List<Bpl.Variable> implInParams,
-      Bpl.Variable/*?*/ resultVariable) {
-      Contract.Requires(f.Formals.Count <= implInParams.Count);
-
-      //generating class post-conditions
-      foreach (var en in f.Ens) {
-        builder.Add(TrAssumeCmdWithDependencies(etran, f.tok, en.E, "overridden function ensures clause"));
-      }
-
-      //generating assume C.F(ins) == out, if a result variable was given
-      if (resultVariable != null) {
-        var funcIdC = new FunctionCall(new Bpl.IdentifierExpr(f.tok, f.FullSanitizedName, TrType(f.ResultType)));
-        var argsC = new List<Bpl.Expr>();
-
-        // add type arguments
-        argsC.AddRange(GetTypeArguments(f, null).ConvertAll(TypeToTy));
-
-        // add fuel arguments
-        if (f.IsFuelAware()) {
-          argsC.Add(etran.layerInterCluster.GetFunctionFuel(f));
-        }
-
-        if (f.IsOpaque || f.IsMadeImplicitlyOpaque(options)) {
-          argsC.Add(GetRevealConstant(f));
-        }
-
-        // add heap arguments
-        if (f is TwoStateFunction) {
-          argsC.Add(etran.Old.HeapExpr);
-        }
-        if (f.ReadsHeap) {
-          argsC.Add(etran.HeapExpr);
-        }
-
-        argsC.AddRange(implInParams.Select(var => new Bpl.IdentifierExpr(f.tok, var)));
-
-        var funcExpC = new Bpl.NAryExpr(f.tok, funcIdC, argsC);
-        var resultVar = new Bpl.IdentifierExpr(resultVariable.tok, resultVariable);
-        builder.Add(TrAssumeCmd(f.tok, Bpl.Expr.Eq(funcExpC, resultVar)));
-      }
-
-      //generating trait post-conditions with class variables
-      FunctionCallSubstituter sub = null;
-      foreach (var en in f.OverriddenFunction.Ens) {
-        sub ??= new FunctionCallSubstituter(substMap, typeMap, (TraitDecl)f.OverriddenFunction.EnclosingClass, (ClassLikeDecl)f.EnclosingClass);
-        foreach (var s in TrSplitExpr(sub.Substitute(en.E), etran, false, out _).Where(s => s.IsChecked)) {
-          builder.Add(Assert(f.tok, s.E, new PODesc.FunctionContractOverride(true)));
-        }
-      }
-    }
-
     /// <summary>
     /// Return type arguments for function "f", where any type parameters are in terms of
     /// the context of "overridingFunction ?? f".
@@ -2017,83 +1963,6 @@ namespace Microsoft.Dafny {
       }
       tyargs.AddRange((overridingFunction ?? f).TypeArgs.ConvertAll(tp => new UserDefinedType(tp.tok, tp)));
       return tyargs;
-    }
-
-    private void AddFunctionOverrideSubsetChk(Function func, BoogieStmtListBuilder builder, ExpressionTranslator etran, List<Variable> localVariables,
-      Dictionary<IVariable, Expression> substMap,
-      Dictionary<TypeParameter, Type> typeMap) {
-      //getting framePrime
-      List<FrameExpression> traitFrameExps = new List<FrameExpression>();
-      FunctionCallSubstituter sub = null;
-      foreach (var e in func.OverriddenFunction.Reads.Expressions) {
-        sub ??= new FunctionCallSubstituter(substMap, typeMap, (TraitDecl)func.OverriddenFunction.EnclosingClass, (ClassLikeDecl)func.EnclosingClass);
-        var newE = sub.Substitute(e.E);
-        FrameExpression fe = new FrameExpression(e.tok, newE, e.FieldName);
-        traitFrameExps.Add(fe);
-      }
-
-      QKeyValue kv = etran.TrAttributes(func.Attributes, null);
-
-      IToken tok = func.tok;
-      // Declare a local variable $_ReadsFrame: <alpha>[ref, Field alpha]bool
-      Bpl.IdentifierExpr traitFrame = etran.ReadsFrame(func.OverriddenFunction.tok);  // this is a throw-away expression, used only to extract the type and name of the $_ReadsFrame variable
-      traitFrame.Name = func.EnclosingClass.Name + "_" + traitFrame.Name;
-      Contract.Assert(traitFrame.Type != null);  // follows from the postcondition of ReadsFrame
-      Bpl.LocalVariable frame = new Bpl.LocalVariable(tok, new Bpl.TypedIdent(tok, null ?? traitFrame.Name, traitFrame.Type));
-      localVariables.Add(frame);
-      // $_ReadsFrame := (lambda<alpha> $o: ref, $f: Field alpha :: $o != null && $Heap[$o,alloc] ==> ($o,$f) in Modifies/Reads-Clause);
-      Bpl.TypeVariable alpha = new Bpl.TypeVariable(tok, "alpha");
-      Bpl.BoundVariable oVar = new Bpl.BoundVariable(tok, new Bpl.TypedIdent(tok, "$o", predef.RefType));
-      Bpl.IdentifierExpr o = new Bpl.IdentifierExpr(tok, oVar);
-      Bpl.BoundVariable fVar = new Bpl.BoundVariable(tok, new Bpl.TypedIdent(tok, "$f", predef.FieldName(tok, alpha)));
-      Bpl.IdentifierExpr f = new Bpl.IdentifierExpr(tok, fVar);
-      Bpl.Expr ante = Bpl.Expr.And(Bpl.Expr.Neq(o, predef.Null), etran.IsAlloced(tok, o));
-      Bpl.Expr consequent = InRWClause(tok, o, f, traitFrameExps, etran, null, null);
-      Bpl.Expr lambda = new Bpl.LambdaExpr(tok, new List<TypeVariable> { alpha }, new List<Variable> { oVar, fVar }, null,
-                                           Bpl.Expr.Imp(ante, consequent));
-
-      //to initialize $_ReadsFrame variable to Frame'
-      builder.Add(Bpl.Cmd.SimpleAssign(tok, new Bpl.IdentifierExpr(tok, frame), lambda));
-
-      // emit: assert (forall<alpha> o: ref, f: Field alpha :: o != null && $Heap[o,alloc] && (o,f) in subFrame ==> $_ReadsFrame[o,f]);
-      Bpl.Expr oInCallee = InRWClause(tok, o, f, func.Reads.Expressions, etran, null, null);
-      Bpl.Expr consequent2 = InRWClause(tok, o, f, traitFrameExps, etran, null, null);
-      Bpl.Expr q = new Bpl.ForallExpr(tok, new List<TypeVariable> { alpha }, new List<Variable> { oVar, fVar },
-                                      Bpl.Expr.Imp(Bpl.Expr.And(ante, oInCallee), consequent2));
-      builder.Add(Assert(tok, q, new PODesc.TraitFrame(func.WhatKind, false), kv));
-    }
-
-    private void AddFunctionOverrideReqsChk(Function f, BoogieStmtListBuilder builder, ExpressionTranslator etran,
-      Dictionary<IVariable, Expression> substMap,
-      Dictionary<TypeParameter, Type> typeMap) {
-      Contract.Requires(f != null);
-      Contract.Requires(builder != null);
-      Contract.Requires(etran != null);
-      Contract.Requires(substMap != null);
-      //generating trait pre-conditions with class variables
-      FunctionCallSubstituter sub = null;
-      foreach (var req in f.OverriddenFunction.Req) {
-        sub ??= new FunctionCallSubstituter(substMap, typeMap, (TraitDecl)f.OverriddenFunction.EnclosingClass, (ClassLikeDecl)f.EnclosingClass);
-        builder.Add(TrAssumeCmdWithDependencies(etran, f.tok, sub.Substitute(req.E), "overridden function requires clause"));
-      }
-      //generating class pre-conditions
-      foreach (var s in f.Req.SelectMany(req => TrSplitExpr(req.E, etran, false, out _).Where(s => s.IsChecked))) {
-        builder.Add(Assert(f.tok, s.E, new PODesc.FunctionContractOverride(false)));
-      }
-    }
-
-    private void HavocMethodFrameLocations(Method m, BoogieStmtListBuilder builder, ExpressionTranslator etran, List<Variable> localVariables) {
-      Contract.Requires(m != null);
-      Contract.Requires(m.EnclosingClass != null && m.EnclosingClass is ClassLikeDecl);
-
-      // play havoc with the heap according to the modifies clause
-      builder.Add(new Bpl.HavocCmd(m.tok, new List<Bpl.IdentifierExpr> { etran.HeapCastToIdentifierExpr }));
-      // assume the usual two-state boilerplate information
-      foreach (BoilerplateTriple tri in GetTwoStateBoilerplate(m.tok, m.Mod.Expressions, m.IsGhost, m.AllowsAllocation, etran.Old, etran, etran.Old)) {
-        if (tri.IsFree) {
-          builder.Add(TrAssumeCmd(m.tok, tri.Expr));
-        }
-      }
     }
 
     private void InsertChecksum(DatatypeDecl d, Bpl.Declaration decl) {
@@ -2343,140 +2212,6 @@ namespace Microsoft.Dafny {
       return false;
     }
 
-    /// <summary>
-    /// Generates:
-    ///   axiom (forall s, h0: HeapType, h1: HeapType, formals... ::
-    ///        { IsHeapAnchor(h0), HeapSucc(h0,h1), F(s,h1,formals) }
-    ///        heaps are well-formed and [formals are allocated AND]
-    ///        IsHeapAnchor(h0) AND HeapSucc(h0,h1)
-    ///        AND
-    ///        (forall(alpha) o: ref, f: Field alpha ::
-    ///            o != null [AND h0[o,alloc] AND]  // note that HeapSucc(h0,h1) && h0[o,alloc] ==> h1[o,alloc]
-    ///            o in reads clause of formals in h0
-    ///            IMPLIES h0[o,f] == h1[o,f])
-    ///        IMPLIES
-    ///        F(s,h0,formals) == F(s,h1,formals)
-    ///      );
-    /// Expressions in [...] are omitted if
-    ///   - /allocated:0, or
-    ///   - /allocated:1, or
-    ///   - /allocated:3, except if "reads" clause is "*" of if the function is a two-state function;
-    /// see comments in AddArrowTypeAxioms
-    /// Also, with /allocated:3, the frame axiom is omitted altogether if the (one-state) function has an
-    /// empty "reads" clause (because then the function doesn't take a heap argument at all).
-    /// </summary>
-    void AddFrameAxiom(Function f) {
-      Contract.Requires(f != null);
-      Contract.Requires(sink != null && predef != null);
-
-      var comment = "frame axiom for " + f.FullSanitizedName;
-      // This is the general case
-      Bpl.Expr prevH = null;
-      Bpl.BoundVariable prevHVar = null;
-      Bpl.Expr reveal = null;
-      Bpl.BoundVariable revealVar = null;
-      if (f.IsOpaque || f.IsMadeImplicitlyOpaque(options)) {
-        revealVar = BplBoundVar("$reveal", Bpl.Type.Bool, out reveal);
-      }
-      if (f is TwoStateFunction) {
-        // The previous-heap argument is the same for both function arguments.  That is,
-        // the frame axiom says nothing about functions invoked with different previous heaps.
-        prevHVar = BplBoundVar("$prevHeap", predef.HeapType, out prevH);
-      }
-      Bpl.Expr h0; var h0Var = BplBoundVar("$h0", predef.HeapType, out h0);
-      Bpl.Expr h1; var h1Var = BplBoundVar("$h1", predef.HeapType, out h1);
-
-      var etran0 = new ExpressionTranslator(this, predef, h0);
-      var etran1 = new ExpressionTranslator(this, predef, h1);
-
-      Bpl.Expr wellFormed = Bpl.Expr.And(
-        FunctionCall(f.tok, BuiltinFunction.IsGoodHeap, null, etran0.HeapExpr),
-        FunctionCall(f.tok, BuiltinFunction.IsGoodHeap, null, etran1.HeapExpr));
-
-      Bpl.TypeVariable alpha = new Bpl.TypeVariable(f.tok, "alpha");
-      Bpl.Expr o; var oVar = BplBoundVar("$o", predef.RefType, out o);
-      Bpl.Expr field; var fieldVar = BplBoundVar("$f", predef.FieldName(f.tok, alpha), out field);
-      Bpl.Expr oNotNull = Bpl.Expr.Neq(o, predef.Null);
-      Bpl.Expr oNotNullAlloced = oNotNull;
-      Bpl.Expr unchanged = Bpl.Expr.Eq(ReadHeap(f.tok, h0, o, field, alpha), ReadHeap(f.tok, h1, o, field, alpha));
-
-      Bpl.Expr h0IsHeapAnchor = FunctionCall(h0.tok, BuiltinFunction.IsHeapAnchor, null, h0);
-      Bpl.Expr heapSucc = HeapSucc(h0, h1);
-      Bpl.Expr r0 = InRWClause(f.tok, o, field, f.Reads.Expressions, etran0, null, null);
-      Bpl.Expr q0 = new Bpl.ForallExpr(f.tok, new List<TypeVariable> { alpha }, new List<Variable> { oVar, fieldVar },
-        Bpl.Expr.Imp(Bpl.Expr.And(oNotNullAlloced, r0), unchanged));
-
-      List<Bpl.Expr> tyexprs;
-      var bvars = MkTyParamBinders(GetTypeParams(f), out tyexprs);
-      var f0args = new List<Bpl.Expr>(tyexprs);
-      var f1args = new List<Bpl.Expr>(tyexprs);
-      var f0argsCanCall = new List<Bpl.Expr>(tyexprs);
-      var f1argsCanCall = new List<Bpl.Expr>(tyexprs);
-      if (f.IsFuelAware()) {
-        Bpl.Expr s; var sV = BplBoundVar("$ly", predef.LayerType, out s);
-        bvars.Add(sV);
-        f0args.Add(s); f1args.Add(s);  // but don't add to f0argsCanCall or f1argsCanCall
-      }
-
-      if (reveal != null) {
-        bvars.Add(revealVar);
-        f0args.Add(reveal); f1args.Add(reveal);
-      }
-
-      if (prevH != null) {
-        bvars.Add(prevHVar);
-        f0args.Add(prevH); f1args.Add(prevH); f0argsCanCall.Add(prevH); f1argsCanCall.Add(prevH);
-      }
-      bvars.Add(h0Var); bvars.Add(h1Var);
-      f0args.Add(h0); f1args.Add(h1); f0argsCanCall.Add(h0); f1argsCanCall.Add(h1);
-
-      var useAlloc = f.Reads.Expressions.Exists(fe => fe.E is WildcardExpr) ? ISALLOC : NOALLOC;
-      if (!f.IsStatic) {
-        Bpl.Expr th; var thVar = BplBoundVar("this", TrReceiverType(f), out th);
-        bvars.Add(thVar);
-        f0args.Add(th); f1args.Add(th); f0argsCanCall.Add(th); f1argsCanCall.Add(th);
-
-        Type thisType = ModuleResolver.GetReceiverType(f.tok, f);
-        Bpl.Expr wh = Bpl.Expr.And(ReceiverNotNull(th), GetWhereClause(f.tok, th, thisType, etran0, useAlloc));
-        wellFormed = Bpl.Expr.And(wellFormed, wh);
-      }
-
-      // (formalsAreWellFormed[h0] || canCallF(h0,...)) && (formalsAreWellFormed[h1] || canCallF(h1,...))
-      Bpl.Expr fwf0 = Bpl.Expr.True;
-      Bpl.Expr fwf1 = Bpl.Expr.True;
-      foreach (Formal p in f.Formals) {
-        Bpl.BoundVariable bv = new Bpl.BoundVariable(p.tok, new Bpl.TypedIdent(p.tok, p.AssignUniqueName(f.IdGenerator), TrType(p.Type)));
-        bvars.Add(bv);
-        Bpl.Expr formal = new Bpl.IdentifierExpr(p.tok, bv);
-        f0args.Add(formal); f1args.Add(formal); f0argsCanCall.Add(formal); f1argsCanCall.Add(formal);
-        Bpl.Expr wh = GetWhereClause(p.tok, formal, p.Type, etran0, useAlloc);
-        if (wh != null) { fwf0 = Bpl.Expr.And(fwf0, wh); }
-      }
-      var canCall = new Bpl.FunctionCall(new Bpl.IdentifierExpr(f.tok, f.FullSanitizedName + "#canCall", Bpl.Type.Bool));
-      wellFormed = Bpl.Expr.And(wellFormed, Bpl.Expr.And(
-        Bpl.Expr.Or(new Bpl.NAryExpr(f.tok, canCall, f0argsCanCall), fwf0),
-        Bpl.Expr.Or(new Bpl.NAryExpr(f.tok, canCall, f1argsCanCall), fwf1)));
-
-      /*
-      DR: I conjecture that this should be enough,
-          as the requires is preserved when the frame is:
-
-      wellFormed = Bpl.Expr.And(wellFormed,
-        Bpl.Expr.Or(new Bpl.NAryExpr(f.tok, canCall, f0argsCanCall), fwf0));
-      */
-
-      var fn = new Bpl.FunctionCall(new Bpl.IdentifierExpr(f.tok, f.FullSanitizedName, TrType(f.ResultType)));
-      var F0 = new Bpl.NAryExpr(f.tok, fn, f0args);
-      var F1 = new Bpl.NAryExpr(f.tok, fn, f1args);
-      var eq = Bpl.Expr.Eq(F0, F1);
-      var tr = new Bpl.Trigger(f.tok, true, new List<Bpl.Expr> { h0IsHeapAnchor, heapSucc, F1 });
-
-      var ax = new Bpl.ForallExpr(f.tok, new List<Bpl.TypeVariable>(), bvars, null, tr,
-        Bpl.Expr.Imp(Bpl.Expr.And(wellFormed, Bpl.Expr.And(h0IsHeapAnchor, heapSucc)),
-        Bpl.Expr.Imp(q0, eq)));
-      sink.AddTopLevelDeclaration(new Bpl.Axiom(f.tok, ax, comment));
-    }
-
     Bpl.Expr InRWClause(IToken tok, Bpl.Expr o, Bpl.Expr f, List<FrameExpression> rw, ExpressionTranslator etran,
                         Expression receiverReplacement, Dictionary<IVariable, Expression> substMap) {
       Contract.Requires(tok != null);
@@ -2578,238 +2313,6 @@ namespace Microsoft.Dafny {
       return disjunction;
     }
 
-    void AddWellformednessCheck(RedirectingTypeDecl decl) {
-      Contract.Requires(decl != null);
-      Contract.Requires(sink != null && predef != null);
-      Contract.Requires(currentModule == null && codeContext == null && isAllocContext == null);
-      Contract.Ensures(currentModule == null && codeContext == null && isAllocContext == null);
-
-      proofDependencies.SetCurrentDefinition(MethodVerboseName(decl.Name, MethodTranslationKind.SpecWellformedness));
-
-      if (!InVerificationScope(decl)) {
-        // Checked in other file
-        return;
-      }
-
-      // If there's no constraint, there's nothing to do
-      if (decl.Var == null) {
-        Contract.Assert(decl.Constraint == null);  // there's a constraint only if there's a variable to be constrained
-        Contract.Assert(decl.WitnessKind == SubsetTypeDecl.WKind.CompiledZero);  // a witness makes sense only if there is a constraint
-        Contract.Assert(decl.Witness == null);  // a witness makes sense only if there is a constraint
-        return;
-      }
-      Contract.Assert(decl.Constraint != null);  // follows from the test above and the RedirectingTypeDecl class invariant
-
-      currentModule = decl.Module;
-      codeContext = new CallableWrapper(decl, true);
-      var etran = new ExpressionTranslator(this, predef, decl.tok);
-
-      // parameters of the procedure
-      var inParams = MkTyParamFormals(decl.TypeArgs, true);
-      Bpl.Type varType = TrType(decl.Var.Type);
-      Bpl.Expr wh = GetWhereClause(decl.Var.tok, new Bpl.IdentifierExpr(decl.Var.tok, decl.Var.AssignUniqueName(decl.IdGenerator), varType), decl.Var.Type, etran, NOALLOC);
-      inParams.Add(new Bpl.Formal(decl.Var.tok, new Bpl.TypedIdent(decl.Var.tok, decl.Var.AssignUniqueName(decl.IdGenerator), varType, wh), true));
-
-      // the procedure itself
-      var req = new List<Bpl.Requires>();
-      // free requires mh == ModuleContextHeight && fh == TypeContextHeight;
-      req.Add(Requires(decl.tok, true, etran.HeightContext(decl), null, null, null));
-      // modifies $Heap
-      var mod = new List<Bpl.IdentifierExpr> {
-        etran.HeapCastToIdentifierExpr,
-      };
-      var name = MethodName(decl, MethodTranslationKind.SpecWellformedness);
-      var proc = new Bpl.Procedure(decl.tok, name, new List<Bpl.TypeVariable>(),
-        inParams, new List<Variable>(),
-        false, req, mod, new List<Bpl.Ensures>(), etran.TrAttributes(decl.Attributes, null));
-      AddVerboseName(proc, decl.Name, MethodTranslationKind.SpecWellformedness);
-      sink.AddTopLevelDeclaration(proc);
-
-      // TODO: Can a checksum be inserted here?
-
-      Contract.Assert(proc.InParams.Count == inParams.Count);
-      // Changed the next line to strip from inParams instead of proc.InParams
-      // They should be the same, but hence the added contract
-      var implInParams = Bpl.Formal.StripWhereClauses(inParams);
-      var locals = new List<Variable>();
-      var builder = new BoogieStmtListBuilder(this, options);
-      builder.Add(new CommentCmd(string.Format("AddWellformednessCheck for {0} {1}", decl.WhatKind, decl)));
-      builder.AddCaptureState(decl.tok, false, "initial state");
-      isAllocContext = new IsAllocContext(options, true);
-
-      DefineFrame(decl.tok, etran.ReadsFrame(decl.tok), new List<FrameExpression>(), builder, locals, null);
-
-      // some initialization stuff;  // This is collected in builderInitializationArea
-      // define frame;
-      // if (*) {
-      //   // The following is collected in constraintCheckBuilder:
-      //   check constraint is well-formed;
-      //   assume constraint;
-      //   do reads checks;
-      // } else {
-      //   // The following is collected in witnessCheckBuilder:
-      //   check witness;
-      // }
-
-      // check well-formedness of the constraint (including termination, and delayed reads checks)
-      var constraintCheckBuilder = new BoogieStmtListBuilder(this, options);
-      var builderInitializationArea = new BoogieStmtListBuilder(this, options);
-      var delayer = new ReadsCheckDelayer(etran, null, locals, builderInitializationArea, constraintCheckBuilder);
-      delayer.DoWithDelayedReadsChecks(false, wfo => {
-        CheckWellformedAndAssume(decl.Constraint, wfo, locals, constraintCheckBuilder, etran, "predicate subtype constraint");
-      });
-
-      // Check that the type is inhabited.
-      // Note, the possible witness in this check should be coordinated with the compiler, so the compiler knows how to do the initialization
-      Expression witnessExpr = null;
-      var witnessCheckBuilder = new BoogieStmtListBuilder(this, options);
-      string witnessString = null;
-      if (decl.Witness != null) {
-        // check well-formedness of the witness expression (including termination, and reads checks)
-        var ghostCodeContext = codeContext;
-        codeContext = decl.WitnessKind == SubsetTypeDecl.WKind.Compiled ? new CallableWrapper(decl, false) : ghostCodeContext;
-        CheckWellformed(decl.Witness, new WFOptions(null, true), locals, witnessCheckBuilder, etran);
-        codeContext = ghostCodeContext;
-        // check that the witness is assignable to the type of the given bound variable
-        if (decl is SubsetTypeDecl) {
-          // Note, for new-types, this has already been checked by CheckWellformed.
-          CheckResultToBeInType(decl.Witness.tok, decl.Witness, decl.Var.Type, locals, witnessCheckBuilder, etran);
-        }
-        // check that the witness expression checks out
-        witnessExpr = Substitute(decl.Constraint, decl.Var, decl.Witness);
-      } else if (decl.WitnessKind == SubsetTypeDecl.WKind.CompiledZero) {
-        var witness = Zero(decl.tok, decl.Var.Type);
-        if (witness == null) {
-          witnessString = "";
-          witnessCheckBuilder.Add(Assert(decl.tok, Bpl.Expr.False, new PODesc.WitnessCheck(witnessString)));
-        } else {
-          // before trying 0 as a witness, check that 0 can be assigned to decl.Var
-          witnessString = Printer.ExprToString(options, witness);
-          CheckResultToBeInType(decl.tok, witness, decl.Var.Type, locals, witnessCheckBuilder, etran, $"trying witness {witnessString}: ");
-          witnessExpr = Substitute(decl.Constraint, decl.Var, witness);
-        }
-      }
-      if (witnessExpr != null) {
-        var witnessCheckTok = decl.Witness != null ? GetToken(decl.Witness) : decl.tok;
-        witnessCheckBuilder.Add(new Bpl.AssumeCmd(witnessCheckTok, CanCallAssumption(witnessExpr, etran)));
-        var witnessCheck = etran.TrExpr(witnessExpr);
-
-        bool splitHappened;
-        var ss = TrSplitExpr(witnessExpr, etran, true, out splitHappened);
-        var desc = new PODesc.WitnessCheck(witnessString);
-        if (!splitHappened) {
-          witnessCheckBuilder.Add(Assert(witnessCheckTok, etran.TrExpr(witnessExpr), desc));
-        } else {
-          foreach (var split in ss) {
-            if (split.IsChecked) {
-              var tok = witnessCheckTok is IToken t ? new NestedToken(t, split.Tok) : witnessCheckTok;
-              witnessCheckBuilder.Add(AssertNS(tok, split.E, desc));
-            }
-          }
-        }
-      }
-
-      builder.Add(new Bpl.IfCmd(decl.tok, null, constraintCheckBuilder.Collect(decl.tok), null, witnessCheckBuilder.Collect(decl.tok)));
-
-      var s0 = builderInitializationArea.Collect(decl.tok);
-      var s1 = builder.Collect(decl.tok);
-      var implBody = new StmtList(new List<BigBlock>(s0.BigBlocks.Concat(s1.BigBlocks)), decl.tok);
-
-      if (EmitImplementation(decl.Attributes)) {
-        // emit the impl only when there are proof obligations.
-        QKeyValue kv = etran.TrAttributes(decl.Attributes, null);
-
-        AddImplementationWithVerboseName(GetToken(decl), proc, implInParams, new List<Variable>(), locals, implBody, kv);
-      }
-
-      // TODO: Should a checksum be inserted here?
-
-      Contract.Assert(currentModule == decl.Module);
-      Contract.Assert(CodeContextWrapper.Unwrap(codeContext) == decl);
-      isAllocContext = null;
-      Reset();
-    }
-
-    void AddWellformednessCheck(ConstantField decl) {
-      Contract.Requires(decl != null);
-      Contract.Requires(sink != null && predef != null);
-      Contract.Requires(currentModule == null && codeContext == null && isAllocContext == null && fuelContext == null);
-      Contract.Ensures(currentModule == null && codeContext == null && isAllocContext == null && fuelContext == null);
-
-      proofDependencies.SetCurrentDefinition(MethodVerboseName(decl.FullDafnyName, MethodTranslationKind.SpecWellformedness));
-      if (!InVerificationScope(decl)) {
-        // Checked in other file
-        return;
-      }
-
-      // If there's no RHS, there's nothing to do
-      if (decl.Rhs == null) {
-        return;
-      }
-
-      currentModule = decl.EnclosingModule;
-      codeContext = decl;
-      fuelContext = FuelSetting.NewFuelContext(decl);
-      var etran = new ExpressionTranslator(this, predef, decl.tok);
-
-      // parameters of the procedure
-      List<Variable> inParams = MkTyParamFormals(GetTypeParams(decl.EnclosingClass), true);
-      if (!decl.IsStatic) {
-        var receiverType = ModuleResolver.GetThisType(decl.tok, (TopLevelDeclWithMembers)decl.EnclosingClass);
-        Contract.Assert(VisibleInScope(receiverType));
-
-        var th = new Bpl.IdentifierExpr(decl.tok, "this", TrReceiverType(decl));
-        var wh = Bpl.Expr.And(
-          ReceiverNotNull(th),
-          etran.GoodRef(decl.tok, th, receiverType));
-        // for class constructors, the receiver is encoded as an output parameter
-        var thVar = new Bpl.Formal(decl.tok, new Bpl.TypedIdent(decl.tok, "this", TrReceiverType(decl), wh), true);
-        inParams.Add(thVar);
-      }
-
-      // the procedure itself
-      var req = new List<Bpl.Requires>();
-      // free requires mh == ModuleContextHeight && fh == TypeContextHeight;
-      req.Add(Requires(decl.tok, true, etran.HeightContext(decl), null, null, null));
-      var heapVar = new Bpl.IdentifierExpr(decl.tok, "$Heap", false);
-      var varlist = new List<Bpl.IdentifierExpr> { heapVar };
-      var name = MethodName(decl, MethodTranslationKind.SpecWellformedness);
-      var proc = new Bpl.Procedure(decl.tok, name, new List<Bpl.TypeVariable>(),
-        inParams, new List<Variable>(),
-        false, req, varlist, new List<Bpl.Ensures>(), etran.TrAttributes(decl.Attributes, null));
-      AddVerboseName(proc, decl.FullDafnyName, MethodTranslationKind.SpecWellformedness);
-      sink.AddTopLevelDeclaration(proc);
-
-      var implInParams = Bpl.Formal.StripWhereClauses(inParams);
-      var locals = new List<Variable>();
-      var builder = new BoogieStmtListBuilder(this, options);
-      builder.Add(new CommentCmd(string.Format("AddWellformednessCheck for {0} {1}", decl.WhatKind, decl)));
-      builder.AddCaptureState(decl.tok, false, "initial state");
-      isAllocContext = new IsAllocContext(options, true);
-
-      DefineFrame(decl.tok, etran.ReadsFrame(decl.tok), new List<FrameExpression>(), builder, locals, null);
-
-      // check well-formedness of the RHS expression
-      CheckWellformed(decl.Rhs, new WFOptions(null, true), locals, builder, etran);
-      builder.Add(new Bpl.AssumeCmd(decl.Rhs.tok, CanCallAssumption(decl.Rhs, etran)));
-      CheckSubrange(decl.Rhs.tok, etran.TrExpr(decl.Rhs), decl.Rhs.Type, decl.Type, builder);
-
-      if (EmitImplementation(decl.Attributes)) {
-        // emit the impl only when there are proof obligations.
-        QKeyValue kv = etran.TrAttributes(decl.Attributes, null);
-        var implBody = builder.Collect(decl.tok);
-
-        AddImplementationWithVerboseName(GetToken(decl), proc, implInParams,
-          new List<Variable>(), locals, implBody, kv);
-      }
-
-      Contract.Assert(currentModule == decl.EnclosingModule);
-      Contract.Assert(codeContext == decl);
-      isAllocContext = null;
-      fuelContext = null;
-      Reset();
-    }
-
     void AddWellformednessCheck(DatatypeCtor ctor) {
       Contract.Requires(ctor != null);
       Contract.Requires(sink != null && predef != null);
@@ -2866,7 +2369,7 @@ namespace Microsoft.Dafny {
       foreach (var formal in ctor.Formals.Where(formal => formal.DefaultValue != null)) {
         var e = formal.DefaultValue;
         CheckWellformed(e, new WFOptions(null, true, false, true), locals, builder, etran);
-        builder.Add(new Bpl.AssumeCmd(e.tok, CanCallAssumption(e, etran)));
+        builder.Add(new Bpl.AssumeCmd(e.tok, etran.CanCallAssumption(e)));
         CheckSubrange(e.tok, etran.TrExpr(e), e.Type, formal.Type, builder);
       }
 
@@ -2951,341 +2454,6 @@ namespace Microsoft.Dafny {
 
       Bpl.IdentifierExpr id = new Bpl.IdentifierExpr(tok, ctor.FullName, predef.DatatypeType);
       return new Bpl.NAryExpr(tok, new Bpl.FunctionCall(id), args);
-    }
-
-    Bpl.Expr CanCallAssumption(Expression expr, ExpressionTranslator etran) {
-      Contract.Requires(expr != null);
-      Contract.Requires(etran != null);
-      Contract.Requires(predef != null);
-      Contract.Ensures(Contract.Result<Bpl.Expr>() != null);
-
-      if (expr is LiteralExpr || expr is ThisExpr || expr is IdentifierExpr || expr is WildcardExpr || expr is BoogieWrapper) {
-        return Bpl.Expr.True;
-      } else if (expr is DisplayExpression) {
-        DisplayExpression e = (DisplayExpression)expr;
-        return CanCallAssumption(e.Elements, etran);
-      } else if (expr is MapDisplayExpr) {
-        MapDisplayExpr e = (MapDisplayExpr)expr;
-        List<Expression> l = new List<Expression>();
-        foreach (ExpressionPair p in e.Elements) {
-          l.Add(p.A); l.Add(p.B);
-        }
-        return CanCallAssumption(l, etran);
-      } else if (expr is MemberSelectExpr) {
-        MemberSelectExpr e = (MemberSelectExpr)expr;
-        var r = CanCallAssumption(e.Obj, etran);
-        if (e.Member is DatatypeDestructor) {
-          var dtor = (DatatypeDestructor)e.Member;
-          if (dtor.EnclosingCtors.Count == dtor.EnclosingCtors[0].EnclosingDatatype.Ctors.Count) {
-            // Every constructor has this destructor; might as well assume that here.
-            var correctConstructor = BplOr(dtor.EnclosingCtors.ConvertAll(
-              ctor => FunctionCall(e.tok, ctor.QueryField.FullSanitizedName, Bpl.Type.Bool, etran.TrExpr(e.Obj))));
-            r = BplAnd(r, correctConstructor);
-          }
-        }
-        return r;
-      } else if (expr is SeqSelectExpr) {
-        SeqSelectExpr e = (SeqSelectExpr)expr;
-        Bpl.Expr total = CanCallAssumption(e.Seq, etran);
-        if (e.E0 != null) {
-          total = BplAnd(total, CanCallAssumption(e.E0, etran));
-        }
-        if (e.E1 != null) {
-          total = BplAnd(total, CanCallAssumption(e.E1, etran));
-        }
-        return total;
-      } else if (expr is MultiSelectExpr) {
-        MultiSelectExpr e = (MultiSelectExpr)expr;
-        Bpl.Expr total = CanCallAssumption(e.Array, etran);
-        foreach (Expression idx in e.Indices) {
-          total = BplAnd(total, CanCallAssumption(idx, etran));
-        }
-        return total;
-      } else if (expr is SeqUpdateExpr) {
-        SeqUpdateExpr e = (SeqUpdateExpr)expr;
-        Bpl.Expr total = CanCallAssumption(e.Seq, etran);
-        total = BplAnd(total, CanCallAssumption(e.Index, etran));
-        total = BplAnd(total, CanCallAssumption(e.Value, etran));
-        return total;
-      } else if (expr is ApplyExpr) {
-        ApplyExpr e = (ApplyExpr)expr;
-        return BplAnd(
-          Cons(CanCallAssumption(e.Function, etran),
-          e.Args.ConvertAll(ee => CanCallAssumption(ee, etran))));
-      } else if (expr is FunctionCallExpr) {
-        FunctionCallExpr e = (FunctionCallExpr)expr;
-        Bpl.Expr r = CanCallAssumption(e.Receiver, etran);
-        r = BplAnd(r, CanCallAssumption(e.Args, etran));
-        if (!(e.Function is SpecialFunction)) {
-          // get to assume canCall
-          Bpl.IdentifierExpr canCallFuncID = new Bpl.IdentifierExpr(expr.tok, e.Function.FullSanitizedName + "#canCall", Bpl.Type.Bool);
-          List<Bpl.Expr> args = etran.FunctionInvocationArguments(e, null, null);
-          Bpl.Expr canCallFuncAppl = new Bpl.NAryExpr(GetToken(expr), new Bpl.FunctionCall(canCallFuncID), args);
-          r = BplAnd(r, canCallFuncAppl);
-        }
-        return r;
-      } else if (expr is DatatypeValue) {
-        DatatypeValue dtv = (DatatypeValue)expr;
-        return CanCallAssumption(dtv.Arguments, etran);
-      } else if (expr is SeqConstructionExpr) {
-        var e = (SeqConstructionExpr)expr;
-        return BplAnd(CanCallAssumption(e.N, etran), CanCallAssumption(e.Initializer, etran));
-      } else if (expr is MultiSetFormingExpr) {
-        MultiSetFormingExpr e = (MultiSetFormingExpr)expr;
-        return CanCallAssumption(e.E, etran);
-      } else if (expr is OldExpr) {
-        var e = (OldExpr)expr;
-        return CanCallAssumption(e.E, etran.OldAt(e.AtLabel));
-      } else if (expr is UnchangedExpr) {
-        var e = (UnchangedExpr)expr;
-        Bpl.Expr be = Bpl.Expr.True;
-        foreach (var fe in e.Frame) {
-          be = BplAnd(be, CanCallAssumption(fe.E, etran));
-        }
-        return be;
-      } else if (expr is UnaryExpr) {
-        var e = (UnaryExpr)expr;
-        return CanCallAssumption(e.E, etran);
-      } else if (expr is BinaryExpr) {
-        // The short-circuiting boolean operators &&, ||, and ==> end up duplicating their
-        // left argument. Therefore, we first try to re-associate the expression to make
-        // left arguments smaller.
-        if (ReAssociateToTheRight(ref expr)) {
-          return CanCallAssumption(expr, etran);
-        }
-        var e = (BinaryExpr)expr;
-
-        Bpl.Expr t0 = CanCallAssumption(e.E0, etran);
-        Bpl.Expr t1 = CanCallAssumption(e.E1, etran);
-        switch (e.ResolvedOp) {
-          case BinaryExpr.ResolvedOpcode.And:
-          case BinaryExpr.ResolvedOpcode.Imp:
-            t1 = BplImp(etran.TrExpr(e.E0), t1);
-            break;
-          case BinaryExpr.ResolvedOpcode.Or:
-            t1 = BplImp(Bpl.Expr.Not(etran.TrExpr(e.E0)), t1);
-            break;
-          case BinaryExpr.ResolvedOpcode.EqCommon:
-          case BinaryExpr.ResolvedOpcode.NeqCommon: {
-              Bpl.Expr r = Bpl.Expr.True;
-              var dt = e.E0.Type.AsDatatype;
-              if (dt != null) {
-                var funcID = new Bpl.FunctionCall(new Bpl.IdentifierExpr(expr.tok, "$IsA#" + dt.FullSanitizedName, Bpl.Type.Bool));
-                if (!(e.E0.Resolved is DatatypeValue)) {
-                  r = BplAnd(r, new Bpl.NAryExpr(expr.tok, funcID, new List<Bpl.Expr> { etran.TrExpr(e.E0) }));
-                }
-                if (!(e.E1.Resolved is DatatypeValue)) {
-                  r = BplAnd(r, new Bpl.NAryExpr(expr.tok, funcID, new List<Bpl.Expr> { etran.TrExpr(e.E1) }));
-                }
-              }
-              return BplAnd(r, BplAnd(t0, t1));
-            }
-          case BinaryExpr.ResolvedOpcode.Mul:
-            if (7 <= options.ArithMode) {
-              if (e.E0.Type.IsNumericBased(Type.NumericPersuasion.Int) && !options.DisableNLarith) {
-                // Produce a useful fact about the associativity of multiplication. It is a bit dicey to do as an axiom.
-                // Change (k*A)*B or (A*k)*B into (A*B)*k, where k is a numeric literal
-                var left = e.E0.Resolved as BinaryExpr;
-                if (left != null && left.ResolvedOp == BinaryExpr.ResolvedOpcode.Mul) {
-                  Bpl.Expr r = Bpl.Expr.True;
-                  if (left.E0.Resolved is LiteralExpr) {
-                    // (K*A)*B == (A*B)*k
-                    var y = Expression.CreateMul(Expression.CreateMul(left.E1, e.E1), left.E0);
-                    var eq = Expression.CreateEq(e, y, e.E0.Type);
-                    r = BplAnd(r, etran.TrExpr(eq));
-                  }
-                  if (left.E1.Resolved is LiteralExpr) {
-                    // (A*k)*B == (A*B)*k
-                    var y = Expression.CreateMul(Expression.CreateMul(left.E0, e.E1), left.E1);
-                    var eq = Expression.CreateEq(e, y, e.E0.Type);
-                    r = BplAnd(r, etran.TrExpr(eq));
-                  }
-                  if (r != Bpl.Expr.True) {
-                    return BplAnd(BplAnd(t0, t1), r);
-                  }
-                }
-              }
-            }
-            break;
-          default:
-            break;
-        }
-        return BplAnd(t0, t1);
-      } else if (expr is TernaryExpr) {
-        var e = (TernaryExpr)expr;
-        return BplAnd(CanCallAssumption(e.E0, etran), BplAnd(CanCallAssumption(e.E1, etran), CanCallAssumption(e.E2, etran)));
-
-      } else if (expr is LetExpr) {
-        var e = (LetExpr)expr;
-        if (!e.Exact) {
-          // CanCall[[ var b0,b1 :| RHS(b0,b1,g); Body(b0,b1,g,h) ]] =
-          //   $let$canCall(g) &&
-          //   CanCall[[ Body($let$b0(g), $let$b1(g), h) ]]
-          LetDesugaring(e);  // call LetDesugaring to prepare the desugaring and populate letSuchThatExprInfo with something for e
-          var info = letSuchThatExprInfo[e];
-          // $let$canCall(g)
-          var canCall = info.CanCallFunctionCall(this, etran);
-          Dictionary<IVariable, Expression> substMap = new Dictionary<IVariable, Expression>();
-          foreach (var bv in e.BoundVars) {
-            // create a call to $let$x(g)
-            var args = info.SkolemFunctionArgs(bv, this, etran);
-            var call = new BoogieFunctionCall(bv.tok, info.SkolemFunctionName(bv), info.UsesHeap, info.UsesOldHeap, info.UsesHeapAt, args.Item1, args.Item2);
-            call.Type = bv.Type;
-            substMap.Add(bv, call);
-          }
-          var p = Substitute(e.Body, null, substMap);
-          var cc = BplAnd(canCall, CanCallAssumption(p, etran));
-          return cc;
-        } else {
-          // CanCall[[ var b := RHS(g); Body(b,g,h) ]] =
-          //   CanCall[[ RHS(g) ]] &&
-          //   (var lhs0,lhs1,... := rhs0,rhs1,...;  CanCall[[ Body ]])
-          Bpl.Expr canCallRHS = Bpl.Expr.True;
-          foreach (var rhs in e.RHSs) {
-            canCallRHS = BplAnd(canCallRHS, CanCallAssumption(rhs, etran));
-          }
-
-          var bodyCanCall = CanCallAssumption(e.Body, etran);
-          // We'd like to compute the free variables if "bodyCanCall". It would be nice to use the Boogie
-          // routine Bpl.Expr.ComputeFreeVariables for this purpose. However, calling it requires the Boogie
-          // expression to be resolved. Instead, we do the cheesy thing of computing the set of names of
-          // free variables in "bodyCanCall".
-          var vis = new VariableNameVisitor();
-          vis.Visit(bodyCanCall);
-
-          List<Bpl.Variable> lhssAll;
-          List<Bpl.Expr> rhssAll;
-          etran.TrLetExprPieces(e, out lhssAll, out rhssAll);
-          Contract.Assert(lhssAll.Count == rhssAll.Count);
-
-          // prune lhss,rhss to contain only those pairs where the LHS is used in the body
-          var lhssPruned = new List<Bpl.Variable>();
-          var rhssPruned = new List<Bpl.Expr>();
-          for (var i = 0; i < lhssAll.Count; i++) {
-            var bv = lhssAll[i];
-            if (vis.Names.Contains(bv.Name)) {
-              lhssPruned.Add(bv);
-              rhssPruned.Add(rhssAll[i]);
-            }
-          }
-          Bpl.Expr let = lhssPruned.Count == 0 ? bodyCanCall : new Bpl.LetExpr(e.tok, lhssPruned, rhssPruned, null, bodyCanCall);
-          return BplAnd(canCallRHS, let);
-        }
-
-      } else if (expr is LambdaExpr) {
-        var e = (LambdaExpr)expr;
-
-        var bvarsAndAntecedents = new List<Tuple<Bpl.Variable, Bpl.Expr>>();
-        var varNameGen = CurrentIdGenerator.NestedFreshIdGenerator("$l#");
-
-        Bpl.Expr heap; var hVar = BplBoundVar(varNameGen.FreshId("#heap#"), predef.HeapType, out heap);
-        var et = new ExpressionTranslator(etran, heap);
-
-        Dictionary<IVariable, Expression> subst = new Dictionary<IVariable, Expression>();
-        foreach (var bv in e.BoundVars) {
-          Bpl.Expr ve; var yVar = BplBoundVar(varNameGen.FreshId(string.Format("#{0}#", bv.Name)), TrType(bv.Type), out ve);
-          var wh = GetWhereClause(bv.tok, new Bpl.IdentifierExpr(bv.tok, yVar), bv.Type, et, NOALLOC);
-          bvarsAndAntecedents.Add(Tuple.Create<Bpl.Variable, Bpl.Expr>(yVar, wh));
-          subst[bv] = new BoogieWrapper(ve, bv.Type);
-        }
-
-        var canCall = CanCallAssumption(Substitute(e.Body, null, subst), et);
-        if (e.Range != null) {
-          var range = Substitute(e.Range, null, subst);
-          canCall = BplAnd(CanCallAssumption(range, etran), BplImp(etran.TrExpr(range), canCall));
-        }
-
-        // It's important to add the heap last to "bvarsAndAntecedents", because the heap may occur in the antecedents of
-        // the other variables and BplForallTrim processes the given tuples in order.
-        var goodHeap = FunctionCall(e.tok, BuiltinFunction.IsGoodHeap, null, heap);
-        bvarsAndAntecedents.Add(Tuple.Create<Bpl.Variable, Bpl.Expr>(hVar, goodHeap));
-
-        //TRIG (forall $l#0#heap#0: Heap, $l#0#x#0: int :: true)
-        //TRIG (forall $l#0#heap#0: Heap, $l#0#t#0: DatatypeType :: _module.__default.TMap#canCall(_module._default.TMap$A, _module._default.TMap$B, $l#0#heap#0, $l#0#t#0, f#0))
-        //TRIG (forall $l#4#heap#0: Heap, $l#4#x#0: Box :: _0_Monad.__default.Bind#canCall(Monad._default.Associativity$B, Monad._default.Associativity$C, $l#4#heap#0, Apply1(Monad._default.Associativity$A, #$M$B, f#0, $l#4#heap#0, $l#4#x#0), g#0))
-        return BplForallTrim(bvarsAndAntecedents, null, canCall); // L_TRIGGER
-
-      } else if (expr is ComprehensionExpr) {
-        var e = (ComprehensionExpr)expr;
-        if (e is QuantifierExpr q && q.SplitQuantifier != null) {
-          return CanCallAssumption(q.SplitQuantifierExpression, etran);
-        }
-
-        // Determine the CanCall's for the range and term
-        var canCall = CanCallAssumption(e.Term, etran);
-        if (e.Range != null) {
-          canCall = BplAnd(CanCallAssumption(e.Range, etran), BplImp(etran.TrExpr(e.Range), canCall));
-        }
-        if (expr is MapComprehension mc && mc.IsGeneralMapComprehension) {
-          canCall = BplAnd(canCall, CanCallAssumption(mc.TermLeft, etran));
-
-          // The translation of "map x,y | R(x,y) :: F(x,y) := G(x,y)" makes use of projection
-          // functions project_x,project_y.  These are functions defined here by the following axiom:
-          //     forall x,y :: R(x,y) ==> var x',y' := project_x(F(x,y)),project_y(F(x,y)); R(x',y') && F(x',y') == F(x,y)
-          // that is (without the let expression):
-          //     forall x,y :: R(x,y) ==> R(project_x(F(x,y)), project_y(F(x,y))) && F(project_x(F(x,y)), project_y(F(x,y))) == F(x,y)
-          // The triggers for the quantification are those detected for the given map comprehension, if any.
-          List<Bpl.Variable> bvs;
-          List<Bpl.Expr> args;
-          CreateBoundVariables(mc.BoundVars, out bvs, out args);
-          Contract.Assert(mc.BoundVars.Count == bvs.Count);
-          CreateMapComprehensionProjectionFunctions(mc);
-          Contract.Assert(mc.ProjectionFunctions != null);
-          Contract.Assert(mc.ProjectionFunctions.Count == mc.BoundVars.Count);
-          var substMap = new Dictionary<IVariable, Expression>();
-          for (var i = 0; i < mc.BoundVars.Count; i++) {
-            substMap.Add(mc.BoundVars[i], new BoogieWrapper(args[i], mc.BoundVars[i].Type));
-          }
-          var R = etran.TrExpr(Substitute(mc.Range, null, substMap));
-          var F = etran.TrExpr(Substitute(mc.TermLeft, null, substMap));
-          var trig = TrTrigger(etran, e.Attributes, expr.tok, substMap);
-          substMap = new Dictionary<IVariable, Expression>();
-          for (var i = 0; i < mc.BoundVars.Count; i++) {
-            var p = new Bpl.NAryExpr(GetToken(mc), new Bpl.FunctionCall(mc.ProjectionFunctions[i]), new List<Bpl.Expr> { F });
-            substMap.Add(e.BoundVars[i], new BoogieWrapper(p, e.BoundVars[i].Type));
-          }
-          var Rprime = etran.TrExpr(Substitute(mc.Range, null, substMap));
-          var Fprime = etran.TrExpr(Substitute(mc.TermLeft, null, substMap));
-          var defn = BplForall(bvs, trig, BplImp(R, BplAnd(Rprime, Bpl.Expr.Eq(F, Fprime))));
-          canCall = BplAnd(canCall, defn);
-        }
-        // Create a list of all possible bound variables
-        var bvarsAndAntecedents = etran.TrBoundVariables_SeparateWhereClauses(e.BoundVars);
-        // Produce the quantified CanCall expression, with a suitably reduced set of bound variables
-        var tr = TrTrigger(etran, e.Attributes, expr.tok);
-        return BplForallTrim(bvarsAndAntecedents, tr, canCall);
-
-      } else if (expr is StmtExpr) {
-        var e = (StmtExpr)expr;
-        return CanCallAssumption(e.E, etran);
-      } else if (expr is ITEExpr) {
-        ITEExpr e = (ITEExpr)expr;
-        Bpl.Expr total = CanCallAssumption(e.Test, etran);
-        Bpl.Expr test = etran.TrExpr(e.Test);
-        total = BplAnd(total, BplImp(test, CanCallAssumption(e.Thn, etran)));
-        total = BplAnd(total, BplImp(Bpl.Expr.Not(test), CanCallAssumption(e.Els, etran)));
-        return total;
-      } else if (expr is ConcreteSyntaxExpression) {
-        var e = (ConcreteSyntaxExpression)expr;
-        return CanCallAssumption(e.ResolvedExpression, etran);
-      } else if (expr is NestedMatchExpr nestedMatchExpr) {
-        return CanCallAssumption(nestedMatchExpr.Flattened, etran);
-      } else if (expr is BoogieFunctionCall) {
-        var e = (BoogieFunctionCall)expr;
-        return CanCallAssumption(e.Args, etran);
-      } else if (expr is MatchExpr) {
-        var e = (MatchExpr)expr;
-        var ite = etran.DesugarMatchExpr(e);
-        return CanCallAssumption(ite, etran);
-      } else if (expr is BoxingCastExpr) {
-        var e = (BoxingCastExpr)expr;
-        return CanCallAssumption(e.E, etran);
-      } else if (expr is UnboxingCastExpr) {
-        var e = (UnboxingCastExpr)expr;
-        return CanCallAssumption(e.E, etran);
-      } else {
-        Contract.Assert(false); throw new cce.UnreachableException();  // unexpected expression
-      }
     }
 
     void AddCasePatternVarSubstitutions(CasePattern<BoundVar> pat, Bpl.Expr rhs, Dictionary<IVariable, Expression> substMap) {
@@ -3379,19 +2547,6 @@ namespace Microsoft.Dafny {
       }
     }
 
-    Bpl.Expr/*!*/ CanCallAssumption(List<Expression/*!*/>/*!*/ exprs, ExpressionTranslator/*!*/ etran) {
-      Contract.Requires(etran != null);
-      Contract.Requires(exprs != null);
-      Contract.Ensures(Contract.Result<Bpl.Expr>() != null);
-
-      Bpl.Expr total = Bpl.Expr.True;
-      foreach (Expression e in exprs) {
-        Contract.Assert(e != null);
-        total = BplAnd(total, CanCallAssumption(e, etran));
-      }
-      return total;
-    }
-
     void CheckNonNull(IToken tok, Expression e, BoogieStmtListBuilder builder, ExpressionTranslator etran, Bpl.QKeyValue kv) {
       Contract.Requires(tok != null);
       Contract.Requires(e != null);
@@ -3409,383 +2564,6 @@ namespace Microsoft.Dafny {
         builder.Add(Assert(tok, Bpl.Expr.Neq(etran.TrExpr(e), predef.Null), new PODesc.NonNull("target object"), kv));
       }
     }
-
-    void TrStmt_CheckWellformed(Expression expr, BoogieStmtListBuilder builder, List<Variable> locals, ExpressionTranslator etran, bool subsumption, bool lValueContext = false) {
-      Contract.Requires(expr != null);
-      Contract.Requires(builder != null);
-      Contract.Requires(locals != null);
-      Contract.Requires(etran != null);
-      Contract.Requires(predef != null);
-
-      Bpl.QKeyValue kv;
-      if (subsumption) {
-        kv = null;  // this is the default behavior of Boogie's assert
-      } else {
-        List<object> args = new List<object>();
-        // {:subsumption 0}
-        args.Add(Bpl.Expr.Literal(0));
-        kv = new Bpl.QKeyValue(expr.tok, "subsumption", args, null);
-      }
-      var options = new WFOptions(kv);
-      // Only do reads checks if reads clauses on methods are enabled and the reads clause is not *.
-      // The latter is important to avoid any extra verification cost for backwards compatibility.
-      if (etran.readsFrame != null) {
-        options = options.WithReadsChecks(true);
-      }
-      if (lValueContext) {
-        options = options.WithLValueContext(true);
-      }
-      CheckWellformed(expr, options, locals, builder, etran);
-      builder.Add(TrAssumeCmd(expr.tok, CanCallAssumption(expr, etran)));
-    }
-
-    /// <summary>
-    /// Returns the translation of converting "r", whose Dafny type was "fromType", to a value of type "toType".
-    /// The translation assumes that "r" is known to be a value of type "toType".
-    /// </summary>
-    Bpl.Expr ConvertExpression(IToken tok, Bpl.Expr r, Type fromType, Type toType) {
-      Contract.Requires(tok != null);
-      Contract.Requires(r != null);
-      Contract.Requires(fromType != null);
-      Contract.Requires(toType != null);
-      toType = toType.NormalizeExpand();
-      fromType = fromType.NormalizeExpand();
-      if (fromType.IsNumericBased(Type.NumericPersuasion.Int)) {
-        if (toType.IsNumericBased(Type.NumericPersuasion.Int)) {
-          // do nothing
-        } else if (toType.IsNumericBased(Type.NumericPersuasion.Real)) {
-          r = FunctionCall(tok, BuiltinFunction.IntToReal, null, r);
-        } else if (toType.IsCharType) {
-          r = FunctionCall(tok, BuiltinFunction.CharFromInt, null, r);
-        } else if (toType.IsBitVectorType) {
-          r = IntToBV(tok, r, toType);
-        } else if (toType.IsBigOrdinalType) {
-          r = FunctionCall(tok, "ORD#FromNat", predef.BigOrdinalType, r);
-        } else {
-          Contract.Assert(false, $"No translation implemented from {fromType} to {toType}");
-        }
-        return r;
-      } else if (fromType.IsBitVectorType) {
-        var fromWidth = fromType.AsBitVectorType.Width;
-        if (toType.IsBitVectorType) {
-          // conversion from one bitvector type to another
-          var toWidth = toType.AsBitVectorType.Width;
-          if (fromWidth == toWidth) {
-            // no conversion
-          } else if (fromWidth < toWidth) {
-            var zeros = BplBvLiteralExpr(tok, BaseTypes.BigNum.ZERO, toWidth - fromWidth);
-            if (fromWidth == 0) {
-              r = zeros;
-            } else {
-              var concat = new Bpl.BvConcatExpr(tok, zeros, r);
-              // There's a bug in Boogie that causes a warning to be emitted if a BvConcatExpr is passed as the argument
-              // to $Box, which takes a type argument.  The bug can apparently be worked around by giving an explicit
-              // (and other redudant) type conversion.
-              r = Bpl.Expr.CoerceType(tok, concat, BplBvType(toWidth));
-            }
-          } else if (toWidth == 0) {
-            r = BplBvLiteralExpr(tok, BaseTypes.BigNum.ZERO, toWidth);
-          } else {
-            r = new Bpl.BvExtractExpr(tok, r, toWidth, 0);
-          }
-        } else if (toType.IsNumericBased(Type.NumericPersuasion.Int)) {
-          r = FunctionCall(tok, "nat_from_bv" + fromWidth, Bpl.Type.Int, r);
-        } else if (toType.IsNumericBased(Type.NumericPersuasion.Real)) {
-          r = FunctionCall(tok, "nat_from_bv" + fromWidth, Bpl.Type.Int, r);
-          r = FunctionCall(tok, BuiltinFunction.IntToReal, null, r);
-        } else if (toType.IsCharType) {
-          r = FunctionCall(tok, "nat_from_bv" + fromWidth, Bpl.Type.Int, r);
-          r = FunctionCall(tok, BuiltinFunction.CharFromInt, null, r);
-        } else if (toType.IsBigOrdinalType) {
-          r = FunctionCall(tok, "nat_from_bv" + fromWidth, Bpl.Type.Int, r);
-          r = FunctionCall(tok, "ORD#FromNat", predef.BigOrdinalType, r);
-        } else {
-          Contract.Assert(false, $"No translation implemented from {fromType} to {toType}");
-        }
-        return r;
-      } else if (fromType.IsCharType) {
-        if (toType.IsNumericBased(Type.NumericPersuasion.Int)) {
-          r = FunctionCall(tok, BuiltinFunction.CharToInt, null, r);
-        } else if (toType.IsCharType) {
-          // do nothing
-        } else if (toType.IsNumericBased(Type.NumericPersuasion.Real)) {
-          r = FunctionCall(tok, BuiltinFunction.CharToInt, null, r);
-          r = FunctionCall(tok, BuiltinFunction.IntToReal, null, r);
-        } else if (toType.IsBitVectorType) {
-          r = FunctionCall(tok, BuiltinFunction.CharToInt, null, r);
-          r = IntToBV(tok, r, toType);
-        } else if (toType.IsBigOrdinalType) {
-          r = FunctionCall(tok, BuiltinFunction.CharToInt, null, r);
-          r = FunctionCall(tok, "ORD#FromNat", Bpl.Type.Int, r);
-        } else {
-          Contract.Assert(false, $"No translation implemented from {fromType} to {toType}");
-        }
-        return r;
-      } else if (fromType.IsNumericBased(Type.NumericPersuasion.Real)) {
-        if (toType.IsNumericBased(Type.NumericPersuasion.Real)) {
-          // do nothing
-        } else if (toType.IsNumericBased(Type.NumericPersuasion.Int)) {
-          r = FunctionCall(tok, BuiltinFunction.RealToInt, null, r);
-        } else if (toType.IsBitVectorType) {
-          r = FunctionCall(tok, BuiltinFunction.RealToInt, null, r);
-          r = IntToBV(tok, r, toType);
-        } else if (toType.IsCharType) {
-          r = FunctionCall(tok, BuiltinFunction.RealToInt, null, r);
-          r = FunctionCall(tok, BuiltinFunction.CharFromInt, null, r);
-        } else if (toType.IsBigOrdinalType) {
-          r = FunctionCall(tok, BuiltinFunction.RealToInt, null, r);
-          r = FunctionCall(tok, "ORD#FromNat", Bpl.Type.Int, r);
-        } else {
-          Contract.Assert(false, $"No translation implemented from {fromType} to {toType}");
-        }
-        return r;
-        // "r" now denotes an integer
-      } else if (fromType.IsBigOrdinalType) {
-        if (toType.IsNumericBased(Type.NumericPersuasion.Int)) {
-          r = FunctionCall(tok, "ORD#Offset", Bpl.Type.Int, r);
-        } else if (toType.IsNumericBased(Type.NumericPersuasion.Real)) {
-          r = FunctionCall(tok, "ORD#Offset", Bpl.Type.Int, r);
-          r = FunctionCall(tok, BuiltinFunction.IntToReal, null, r);
-        } else if (toType.IsCharType) {
-          r = FunctionCall(tok, "ORD#Offset", Bpl.Type.Int, r);
-          r = FunctionCall(tok, BuiltinFunction.CharFromInt, null, r);
-        } else if (toType.IsBitVectorType) {
-          r = FunctionCall(tok, "ORD#Offset", Bpl.Type.Int, r);
-          r = IntToBV(tok, r, toType);
-        } else if (toType.IsBigOrdinalType) {
-          // do nothing
-        } else {
-          Contract.Assert(false, $"No translation implemented from {fromType} to {toType}");
-        }
-        return r;
-      } else if (fromType.IsRefType && toType.IsRefType) {
-        return r;
-      } else if (fromType.IsRefType) {
-        Contract.Assert(toType.IsTraitType);
-        return BoxIfNecessary(r.tok, r, fromType);
-      } else if (toType.IsRefType) {
-        Contract.Assert(fromType.IsTraitType);
-        return UnboxIfBoxed(r, toType);
-      } else if (toType.IsTraitType) {
-        // cast to a non-reference trait
-        return BoxIfNecessary(r.tok, r, fromType);
-      } else if (fromType.IsTraitType) {
-        // cast from a non-reference trait
-        return UnboxIfBoxed(r, toType);
-      } else {
-        Contract.Assert(false, $"No translation implemented from {fromType} to {toType}");
-      }
-      return r;
-    }
-
-    private Bpl.Expr IntToBV(IToken tok, Bpl.Expr r, Type toType) {
-      var toWidth = toType.AsBitVectorType.Width;
-      if (RemoveLit(r) is Bpl.LiteralExpr) {
-        Bpl.LiteralExpr e = (Bpl.LiteralExpr)RemoveLit(r);
-        if (e.isBigNum) {
-          var toBound = BaseTypes.BigNum.FromBigInt(BigInteger.One << toWidth);  // 1 << toWidth
-          if (e.asBigNum <= toBound) {
-            return BplBvLiteralExpr(r.tok, e.asBigNum, toType.AsBitVectorType);
-          }
-        }
-      }
-      return FunctionCall(tok, "nat_to_bv" + toWidth, BplBvType(toWidth), r);
-    }
-
-    /// <summary>
-    /// Emit checks that "expr" (which may or may not be a value of type "expr.Type"!) is a value of type "toType".
-    /// </summary>
-    void CheckResultToBeInType(IToken tok, Expression expr, Type toType, List<Bpl.Variable> locals, BoogieStmtListBuilder builder, ExpressionTranslator etran, string errorMsgPrefix = "") {
-      Contract.Requires(tok != null);
-      Contract.Requires(expr != null);
-      Contract.Requires(toType != null);
-      Contract.Requires(builder != null);
-      Contract.Requires(etran != null);
-      Contract.Requires(errorMsgPrefix != null);
-
-      // Lazily create a local variable "o" to hold the value of the from-expression
-      Bpl.IdentifierExpr o = null;
-      void PutSourceIntoLocal() {
-        if (o == null) {
-          var oType = expr.Type.IsCharType ? Type.Int : expr.Type;
-          var oVar = new Bpl.LocalVariable(tok, new Bpl.TypedIdent(tok, CurrentIdGenerator.FreshId("newtype$check#"), TrType(oType)));
-          locals.Add(oVar);
-          o = new Bpl.IdentifierExpr(tok, oVar);
-          var rhs = etran.TrExpr(expr);
-          if (expr.Type.IsCharType) {
-            rhs = FunctionCall(expr.tok, "char#ToInt", Bpl.Type.Int, rhs);
-          }
-          builder.Add(Bpl.Cmd.SimpleAssign(tok, o, rhs));
-        }
-      }
-
-      Contract.Assert(options.Get(CommonOptionBag.GeneralTraits) != CommonOptionBag.GeneralTraitsOptions.Legacy || expr.Type.IsRefType == toType.IsRefType);
-      if (toType.IsRefType) {
-        PutSourceIntoLocal();
-        CheckSubrange(tok, o, expr.Type, toType, builder, errorMsgPrefix);
-        return;
-      } else if (expr.Type.IsTraitType) {
-        PutSourceIntoLocal();
-        CheckSubrange(tok, o, expr.Type, toType, builder, errorMsgPrefix);
-        return;
-      }
-
-      if (expr.Type.IsNumericBased(Type.NumericPersuasion.Real) && !toType.IsNumericBased(Type.NumericPersuasion.Real)) {
-        // this operation is well-formed only if the real-based number represents an integer
-        //   assert Real(Int(o)) == o;
-        PutSourceIntoLocal();
-        Bpl.Expr from = FunctionCall(tok, BuiltinFunction.RealToInt, null, o);
-        Bpl.Expr e = FunctionCall(tok, BuiltinFunction.IntToReal, null, from);
-        e = Bpl.Expr.Binary(tok, Bpl.BinaryOperator.Opcode.Eq, e, o);
-        builder.Add(Assert(tok, e, new PODesc.IsInteger(errorMsgPrefix)));
-      }
-
-      if (expr.Type.IsBigOrdinalType && !toType.IsBigOrdinalType) {
-        PutSourceIntoLocal();
-        Bpl.Expr boundsCheck = FunctionCall(tok, "ORD#IsNat", Bpl.Type.Bool, o);
-        builder.Add(Assert(tok, boundsCheck, new PODesc.ConversionIsNatural(errorMsgPrefix)));
-      }
-
-      if (toType.IsBitVectorType) {
-        var toWidth = toType.AsBitVectorType.Width;
-        var toBound = BaseTypes.BigNum.FromBigInt(BigInteger.One << toWidth);  // 1 << toWidth
-        Bpl.Expr boundsCheck = null;
-        if (expr.Type.IsBitVectorType) {
-          var fromWidth = expr.Type.AsBitVectorType.Width;
-          if (toWidth < fromWidth) {
-            // Check "expr < (1 << toWidth)" in type "fromType" (note that "1 << toWidth" is indeed a value in "fromType")
-            PutSourceIntoLocal();
-            var bound = BplBvLiteralExpr(tok, toBound, expr.Type.AsBitVectorType);
-            boundsCheck = FunctionCall(expr.tok, "lt_bv" + fromWidth, Bpl.Type.Bool, o, bound);
-          }
-        } else if (expr.Type.IsNumericBased(Type.NumericPersuasion.Int) || expr.Type.IsCharType) {
-          // Check "expr < (1 << toWdith)" in type "int"
-          PutSourceIntoLocal();
-          var bound = Bpl.Expr.Literal(toBound);
-          boundsCheck = Bpl.Expr.And(Bpl.Expr.Le(Bpl.Expr.Literal(0), o), Bpl.Expr.Lt(o, bound));
-        } else if (expr.Type.IsNumericBased(Type.NumericPersuasion.Real)) {
-          // Check "Int(expr) < (1 << toWdith)" in type "int"
-          PutSourceIntoLocal();
-          var bound = Bpl.Expr.Literal(toBound);
-          var oi = FunctionCall(tok, BuiltinFunction.RealToInt, null, o);
-          boundsCheck = Bpl.Expr.And(Bpl.Expr.Le(Bpl.Expr.Literal(0), oi), Bpl.Expr.Lt(oi, bound));
-        } else if (expr.Type.IsBigOrdinalType) {
-          var bound = Bpl.Expr.Literal(toBound);
-          var oi = FunctionCall(tok, "ORD#Offset", Bpl.Type.Int, o);
-          boundsCheck = Bpl.Expr.Lt(oi, bound);
-        }
-
-        if (boundsCheck != null) {
-          builder.Add(Assert(tok, boundsCheck, new PODesc.ConversionFit("value", toType, errorMsgPrefix)));
-        }
-      }
-
-      if (toType.IsCharType) {
-        if (expr.Type.IsNumericBased(Type.NumericPersuasion.Int)) {
-          PutSourceIntoLocal();
-          var boundsCheck = FunctionCall(Token.NoToken, BuiltinFunction.IsChar, null, o);
-          builder.Add(Assert(tok, boundsCheck, new PODesc.ConversionFit("value", toType, errorMsgPrefix)));
-        } else if (expr.Type.IsNumericBased(Type.NumericPersuasion.Real)) {
-          PutSourceIntoLocal();
-          var oi = FunctionCall(tok, BuiltinFunction.RealToInt, null, o);
-          var boundsCheck = FunctionCall(Token.NoToken, BuiltinFunction.IsChar, null, oi);
-          builder.Add(Assert(tok, boundsCheck, new PODesc.ConversionFit("real value", toType, errorMsgPrefix)));
-        } else if (expr.Type.IsBitVectorType) {
-          PutSourceIntoLocal();
-          var fromWidth = expr.Type.AsBitVectorType.Width;
-          var toWidth = 16;
-          if (toWidth < fromWidth) {
-            // Check "expr < (1 << toWidth)" in type "fromType" (note that "1 << toWidth" is indeed a value in "fromType")
-            PutSourceIntoLocal();
-            var toBound = BaseTypes.BigNum.FromBigInt(BigInteger.One << toWidth); // 1 << toWidth
-            var bound = BplBvLiteralExpr(tok, toBound, expr.Type.AsBitVectorType);
-            var boundsCheck = FunctionCall(expr.tok, "lt_bv" + fromWidth, Bpl.Type.Bool, o, bound);
-            builder.Add(Assert(tok, boundsCheck, new PODesc.ConversionFit("bit-vector value", toType, errorMsgPrefix)));
-          }
-        } else if (expr.Type.IsBigOrdinalType) {
-          PutSourceIntoLocal();
-          var oi = FunctionCall(tok, "ORD#Offset", Bpl.Type.Int, o);
-          int toWidth = 16;
-          var toBound = BaseTypes.BigNum.FromBigInt(BigInteger.One << toWidth); // 1 << toWidth
-          var bound = Bpl.Expr.Literal(toBound);
-          var boundsCheck = Bpl.Expr.Lt(oi, bound);
-          builder.Add(Assert(tok, boundsCheck, new PODesc.ConversionFit("ORDINAL value", toType, errorMsgPrefix)));
-        }
-      } else if (toType.IsBigOrdinalType) {
-        if (expr.Type.IsNumericBased(Type.NumericPersuasion.Int)) {
-          PutSourceIntoLocal();
-          Bpl.Expr boundsCheck = Bpl.Expr.Le(Bpl.Expr.Literal(0), o);
-          var desc = new PODesc.ConversionPositive("integer", toType, errorMsgPrefix);
-          builder.Add(Assert(tok, boundsCheck, desc));
-        }
-        if (expr.Type.IsNumericBased(Type.NumericPersuasion.Real)) {
-          PutSourceIntoLocal();
-          var oi = FunctionCall(tok, BuiltinFunction.RealToInt, null, o);
-          Bpl.Expr boundsCheck = Bpl.Expr.Le(Bpl.Expr.Literal(0), oi);
-          var desc = new PODesc.ConversionPositive("real", toType, errorMsgPrefix);
-          builder.Add(Assert(tok, boundsCheck, desc));
-        }
-      } else if (toType.IsNumericBased(Type.NumericPersuasion.Int)) {
-        // already checked that BigOrdinal or real inputs are integral
-      } else if (toType.IsNumericBased(Type.NumericPersuasion.Real)) {
-        // already checked that BigOrdinal is integral
-      }
-
-      if (toType.NormalizeExpandKeepConstraints().AsRedirectingType != null) {
-        PutSourceIntoLocal();
-        Bpl.Expr be;
-        if (expr.Type.IsNumericBased() || expr.Type.IsBitVectorType) {
-          be = ConvertExpression(expr.tok, o, expr.Type, toType);
-        } else if (expr.Type.IsCharType) {
-          be = ConvertExpression(expr.tok, o, Dafny.Type.Int, toType);
-        } else if (expr.Type.IsBigOrdinalType) {
-          be = FunctionCall(expr.tok, "ORD#Offset", Bpl.Type.Int, o);
-          be = ConvertExpression(expr.tok, be, Dafny.Type.Int, toType);
-        } else {
-          be = o;
-        }
-        var dafnyType = toType.NormalizeExpand();
-        CheckResultToBeInType_Aux(tok, new BoogieWrapper(be, dafnyType), toType.NormalizeExpandKeepConstraints(), builder, etran, errorMsgPrefix);
-      }
-    }
-
-    void CheckResultToBeInType_Aux(IToken tok, Expression expr, Type toType, BoogieStmtListBuilder builder, ExpressionTranslator etran, string errorMsgPrefix) {
-      Contract.Requires(tok != null);
-      Contract.Requires(expr != null);
-      Contract.Requires(toType != null && toType.AsRedirectingType != null);
-      Contract.Requires(builder != null);
-      Contract.Requires(etran != null);
-      Contract.Requires(errorMsgPrefix != null);
-      // First, check constraints of base types
-      var udt = (UserDefinedType)toType;
-      var rdt = (RedirectingTypeDecl)udt.ResolvedClass;
-      Type baseType;
-      string kind;
-      if (rdt is SubsetTypeDecl) {
-        baseType = ((SubsetTypeDecl)rdt).RhsWithArgument(udt.TypeArgs);
-        kind = "subset type";
-      } else if (rdt is NewtypeDecl) {
-        baseType = ((NewtypeDecl)rdt).BaseType;
-        kind = "newtype";
-      } else {
-        baseType = ((TypeSynonymDecl)rdt).RhsWithArgument(udt.TypeArgs);
-        kind = "type synonym";
-      }
-
-      if (baseType.AsRedirectingType != null) {
-        CheckResultToBeInType_Aux(tok, expr, baseType, builder, etran, errorMsgPrefix);
-      }
-      // Check any constraint defined in 'dd'
-      if (rdt.Var != null) {
-        // TODO: use TrSplitExpr
-        var substMap = new Dictionary<IVariable, Expression>();
-        substMap.Add(rdt.Var, expr);
-        var typeMap = TypeParameter.SubstitutionMap(rdt.TypeArgs, udt.TypeArgs);
-        var constraint = etran.TrExpr(Substitute(rdt.Constraint, null, substMap, typeMap));
-        builder.Add(Assert(tok, constraint, new PODesc.ConversionSatisfiesConstraints(errorMsgPrefix, kind, rdt.Name)));
-      }
-    }
-
 
     void CheckFunctionSelectWF(string what, BoogieStmtListBuilder builder, ExpressionTranslator etran, Expression e, string hint) {
       if (e is MemberSelectExpr sel && sel.Member is Function fn) {
@@ -3856,182 +2634,6 @@ namespace Microsoft.Dafny {
 
     public string RequiresName(Function f) {
       return f.FullSanitizedName + "#requires";
-    }
-
-    public string FunctionHandle(Function f) {
-      Contract.Requires(f != null);
-      string name;
-      if (functionHandles.TryGetValue(f, out name)) {
-        Contract.Assert(name != null);
-      } else {
-        name = f.FullSanitizedName + "#Handle";
-        functionHandles[f] = name;
-        var args = new List<Bpl.Expr>();
-        var vars = MkTyParamBinders(GetTypeParams(f), out args);
-        var argsRequires = new List<Bpl.Expr>(args); // Requires don't have reveal parameters
-        var formals = MkTyParamFormals(GetTypeParams(f), false, true);
-        var tyargs = new List<Bpl.Expr>();
-        foreach (var fm in f.Formals) {
-          tyargs.Add(TypeToTy(fm.Type));
-        }
-        tyargs.Add(TypeToTy(f.ResultType));
-        if (f.IsFuelAware()) {
-          vars.Add(BplBoundVar("$ly", predef.LayerType, out var ly));
-          args.Add(ly);
-          argsRequires.Add(ly);
-          formals.Add(BplFormalVar("$fuel", predef.LayerType, true));
-          AddFuelSuccSynonymAxiom(f, true);
-        }
-        if (f.IsOpaque || f.IsMadeImplicitlyOpaque(options)) {
-          vars.Add(BplBoundVar("$reveal", Boogie.Type.Bool, out var reveal));
-          args.Add(reveal);
-          formals.Add(BplFormalVar("$reveal", Boogie.Type.Bool, true));
-        }
-
-        Func<List<Bpl.Expr>, List<Bpl.Expr>> SnocSelf = x => x;
-        Func<List<Bpl.Expr>, List<Bpl.Expr>> SnocPrevH = x => x;
-        Expression selfExpr;
-        Dictionary<IVariable, Expression> rhs_dict = new Dictionary<IVariable, Expression>();
-        if (f is TwoStateFunction) {
-          // also add previous-heap to the list of fixed arguments of the handle
-          var prevH = BplBoundVar("$prevHeap", predef.HeapType, vars);
-          formals.Add(BplFormalVar("h", predef.HeapType, true));
-          SnocPrevH = xs => Snoc(xs, prevH);
-        }
-        if (f.IsStatic) {
-          selfExpr = null;
-        } else {
-          var selfTy = TrType(UserDefinedType.FromTopLevelDecl(f.tok, f.EnclosingClass));
-          var self = BplBoundVar("$self", selfTy, vars);
-          formals.Add(BplFormalVar("self", selfTy, true));
-          SnocSelf = xs => Snoc(xs, self);
-          var wrapperType = UserDefinedType.FromTopLevelDecl(f.tok, f.EnclosingClass);
-          selfExpr = new BoogieWrapper(self, wrapperType);
-        }
-
-        // F#Handle(Ty, .., Ty, LayerType, ref) : HandleType
-        sink.AddTopLevelDeclaration(
-          new Bpl.Function(f.tok, name, formals, BplFormalVar(null, predef.HandleType, false)));
-
-        var bvars = new List<Bpl.Variable>();
-        var lhs_args = new List<Bpl.Expr>();
-        var rhs_args = new List<Bpl.Expr>();
-        var func_vars = new List<Bpl.Variable>();
-        var func_args = new List<Bpl.Expr>();
-        var boxed_func_args = new List<Bpl.Expr>();
-
-        var idGen = f.IdGenerator.NestedFreshIdGenerator("$fh$");
-        foreach (var fm in f.Formals) {
-          string fm_name = idGen.FreshId("x#");
-          // Box and its [Unbox]args
-          var fe = BplBoundVar(fm_name, predef.BoxType, bvars);
-          lhs_args.Add(fe);
-          var be = UnboxIfBoxed(fe, fm.Type);
-          rhs_args.Add(be);
-
-          rhs_dict[fm] = new BoogieWrapper(be, fm.Type);
-          // args and its [Box]args
-          var arg = BplBoundVar(fm_name, TrType(fm.Type), func_vars);
-          func_args.Add(arg);
-          var boxed = BoxIfUnboxed(arg, fm.Type);
-          boxed_func_args.Add(boxed);
-        }
-
-        var h = BplBoundVar("$heap", predef.HeapType, vars);
-
-        int arity = f.Formals.Count;
-
-        {
-          // Apply(Ty.., F#Handle( Ty1, ..., TyN, Layer, self), Heap, arg1, ..., argN)
-          //   = [Box] F(Ty1, .., TyN, Layer, Heap, self, [Unbox] arg1, .., [Unbox] argN)
-
-          var fhandle = FunctionCall(f.tok, name, predef.HandleType, SnocSelf(SnocPrevH(args)));
-          var lhs = FunctionCall(f.tok, Apply(arity), TrType(f.ResultType),
-            Concat(tyargs, Cons(h, Cons(fhandle, lhs_args))));
-          var args_h = f.ReadsHeap ? Snoc(SnocPrevH(args), h) : args;
-          var rhs = FunctionCall(f.tok, f.FullSanitizedName, TrType(f.ResultType), Concat(SnocSelf(args_h), rhs_args));
-          var rhs_boxed = BoxIfUnboxed(rhs, f.ResultType);
-
-          AddOtherDefinition(GetOrCreateFunction(f), (new Axiom(f.tok,
-            BplForall(Concat(vars, bvars), BplTrigger(lhs), Bpl.Expr.Eq(lhs, rhs_boxed)))));
-        }
-
-        {
-          // As a first approximation, the following axiom is of the form:
-          // Requires(Ty.., F#Handle( Ty1, ..., TyN, Layer, reveal, self), Heap, arg1, ..., argN)
-          //   = F#Requires(Ty1, .., TyN, Layer, Heap, self, [Unbox] arg1, .., [Unbox] argN)
-          // However, .reads ands .requires functions require special attention.
-          // To understand the rationale for these axioms, refer to the section on arrow types of the reference manual.
-          // The requires clause of the .requires function is simply true.
-          // The requires clause of the .reads function checks that the precondtion of the receiving function holds.
-
-          var fhandle = FunctionCall(f.tok, name, predef.HandleType, SnocSelf(SnocPrevH(args)));
-          var lhs = FunctionCall(f.tok, Requires(arity), Bpl.Type.Bool, Concat(tyargs, Cons(h, Cons(fhandle, lhs_args))));
-          Bpl.Expr rhs;
-          if (f.EnclosingClass is ArrowTypeDecl && f.Name == "requires") {
-            AddOtherDefinition(GetOrCreateFunction(f), new Axiom(f.tok,
-                BplForall(Concat(vars, bvars), BplTrigger(lhs), Bpl.Expr.Eq(lhs, Bpl.Expr.True))));
-          } else if (f.EnclosingClass is ArrowTypeDecl && f.Name == "reads") {
-            var args_h = f.ReadsHeap ? Snoc(SnocPrevH(argsRequires), h) : argsRequires;
-            var pre = FunctionCall(f.tok, Requires(arity), Bpl.Type.Bool, Concat(SnocSelf(args_h), lhs_args));
-            AddOtherDefinition(GetOrCreateFunction(f), (new Axiom(f.tok,
-              BplForall(Concat(vars, bvars), BplTrigger(lhs), Bpl.Expr.Eq(lhs, pre)))));
-          } else {
-            var args_h = f.ReadsHeap ? Snoc(SnocPrevH(argsRequires), h) : argsRequires;
-            rhs = FunctionCall(f.tok, RequiresName(f), Bpl.Type.Bool, Concat(SnocSelf(args_h), rhs_args));
-            AddOtherDefinition(GetOrCreateFunction(f), new Axiom(f.tok,
-              BplForall(Concat(vars, bvars), BplTrigger(lhs), Bpl.Expr.Eq(lhs, rhs))));
-          }
-
-
-        }
-
-        {
-          // As a first approximation, the following axiom is of the form:
-          // Reads(Ty.., F#Handle( Ty1, ..., TyN, Layer, self), Heap, arg1, ..., argN)
-          //   =  $Frame_F(args...)
-          // However, .reads ands .requires functions require special attention.
-          // To understand the rationale for these axioms, refer to the section on arrow types of the reference manual.
-          // In both cases, the precondition of the receiving function must be checked before its reads clause can
-          // be referred to.
-
-          var fhandle = FunctionCall(f.tok, name, predef.HandleType, SnocSelf(SnocPrevH(args)));
-          Bpl.Expr lhs_inner = FunctionCall(f.tok, Reads(arity), TrType(program.SystemModuleManager.ObjectSetType()), Concat(tyargs, Cons(h, Cons(fhandle, lhs_args))));
-
-          Bpl.Expr bx; var bxVar = BplBoundVar("$bx", predef.BoxType, out bx);
-          Bpl.Expr unboxBx = FunctionCall(f.tok, BuiltinFunction.Unbox, predef.RefType, bx);
-          Bpl.Expr lhs = Bpl.Expr.SelectTok(f.tok, lhs_inner, bx);
-
-          var et = new ExpressionTranslator(this, predef, h);
-          var rhs = InRWClause_Aux(f.tok, unboxBx, bx, null, f.Reads.Expressions, false, et, selfExpr, rhs_dict);
-
-          if (f.EnclosingClass is ArrowTypeDecl) {
-            var args_h = f.ReadsHeap ? Snoc(SnocPrevH(argsRequires), h) : argsRequires;
-            var precondition = FunctionCall(f.tok, Requires(arity), Bpl.Type.Bool, Concat(SnocSelf(args_h), lhs_args));
-            sink.AddTopLevelDeclaration(new Axiom(f.tok,
-              BplForall(Cons(bxVar, Concat(vars, bvars)), BplTrigger(lhs), Bpl.Expr.Imp(precondition, Bpl.Expr.Eq(lhs, rhs)))));
-          } else {
-            sink.AddTopLevelDeclaration(new Axiom(f.tok,
-              BplForall(Cons(bxVar, Concat(vars, bvars)), BplTrigger(lhs), Bpl.Expr.Eq(lhs, rhs))));
-          }
-        }
-
-        {
-          // F(Ty1, .., TyN, Layer, Heap, self, arg1, .., argN)
-          // = [Unbox]Apply1(Ty.., F#Handle( Ty1, ..., TyN, Layer, self), Heap, [Box]arg1, ..., [Box]argN)
-
-          var fhandle = FunctionCall(f.tok, name, predef.HandleType, SnocSelf(SnocPrevH(args)));
-          var args_h = f.ReadsHeap ? Snoc(SnocPrevH(args), h) : args;
-          var lhs = FunctionCall(f.tok, f.FullSanitizedName, TrType(f.ResultType), Concat(SnocSelf(args_h), func_args));
-          var rhs = FunctionCall(f.tok, Apply(arity), TrType(f.ResultType), Concat(tyargs, Cons(h, Cons(fhandle, boxed_func_args))));
-          var rhs_unboxed = UnboxIfBoxed(rhs, f.ResultType);
-          var tr = BplTriggerHeap(this, f.tok, lhs, f.ReadsHeap ? null : h);
-
-          AddOtherDefinition(GetOrCreateFunction(f), (new Axiom(f.tok,
-            BplForall(Concat(vars, func_vars), tr, Bpl.Expr.Eq(lhs, rhs_unboxed)))));
-        }
-      }
-      return name;
     }
 
     private Expr NewOneHeapExpr(IToken tok) {
@@ -4182,140 +2784,6 @@ namespace Microsoft.Dafny {
         fieldConstants.Add(n, cc);
       }
       return cc;
-    }
-
-    Bpl.Constant GetField(Field f) {
-      Contract.Requires(f != null && f.IsMutable);
-      Contract.Requires(sink != null && predef != null);
-      Contract.Ensures(Contract.Result<Bpl.Constant>() != null);
-
-      Contract.Assert(VisibleInScope(f));
-
-      Bpl.Constant fc;
-      if (fields.TryGetValue(f, out fc)) {
-        Contract.Assert(fc != null);
-      } else {
-        // const f: Field ty;
-        Bpl.Type ty = predef.FieldName(f.tok, TrType(f.Type));
-        fc = new Bpl.Constant(f.tok, new Bpl.TypedIdent(f.tok, f.FullSanitizedName, ty), false);
-        fields.Add(f, fc);
-        // axiom FDim(f) == 0 && FieldOfDecl(C, name) == f &&
-        //       $IsGhostField(f);    // if the field is a ghost field
-        // OR:
-        //       !$IsGhostField(f);    // if the field is not a ghost field
-        Bpl.Expr fdim = Bpl.Expr.Eq(FunctionCall(f.tok, BuiltinFunction.FDim, ty, Bpl.Expr.Ident(fc)), Bpl.Expr.Literal(0));
-        Bpl.Expr declType = Bpl.Expr.Eq(FunctionCall(f.tok, BuiltinFunction.FieldOfDecl, ty, new Bpl.IdentifierExpr(f.tok, GetClass(cce.NonNull(f.EnclosingClass))), new Bpl.IdentifierExpr(f.tok, GetFieldNameFamily(f.Name))), Bpl.Expr.Ident(fc));
-        Bpl.Expr cond = Bpl.Expr.And(fdim, declType);
-        var ig = FunctionCall(f.tok, BuiltinFunction.IsGhostField, ty, Bpl.Expr.Ident(fc));
-        cond = Bpl.Expr.And(cond, f.IsGhost ? ig : Bpl.Expr.Not(ig));
-        Bpl.Axiom ax = new Bpl.Axiom(f.tok, cond);
-        AddOtherDefinition(fc, ax);
-      }
-      return fc;
-    }
-
-
-    Bpl.Function GetReadonlyField(Field f) {
-      Contract.Requires(f != null && !f.IsMutable);
-      Contract.Requires(sink != null && predef != null);
-      Contract.Ensures(Contract.Result<Bpl.Function>() != null);
-
-      Contract.Assert(VisibleInScope(f));
-
-      Bpl.Function ff;
-      if (fieldFunctions.TryGetValue(f, out ff)) {
-        Contract.Assert(ff != null);
-      } else {
-        // Here are some built-in functions defined in "predef" (so there's no need to cache them in "fieldFunctions")
-        if (f.EnclosingClass is ArrayClassDecl && f.Name == "Length") {
-          return predef.ArrayLength;
-        } else if (f.EnclosingClass is ValuetypeDecl { Name: "real" } && f.Name == "Floor") {
-          return predef.RealFloor;
-        } else if (f is SpecialField && !(f is DatatypeDestructor || f.EnclosingClass is TopLevelDeclWithMembers and not ValuetypeDecl)) {
-          if (f.Name is "Keys" or "Values" or "Items") {
-            var fType = f.Type.NormalizeExpand();
-            Contract.Assert(fType is SetType);
-            var setType = (SetType)fType;
-            return f.Name switch {
-              "Keys" => setType.Finite ? predef.MapDomain : predef.IMapDomain,
-              "Values" => setType.Finite ? predef.MapValues : predef.IMapValues,
-              _ => setType.Finite ? predef.MapItems : predef.IMapItems
-            };
-          }
-          if (f.Name == "IsLimit") {
-            return predef.ORDINAL_IsLimit;
-          } else if (f.Name == "IsSucc") {
-            return predef.ORDINAL_IsSucc;
-          } else if (f.Name == "Offset") {
-            return predef.ORDINAL_Offset;
-          } else if (f.Name == "IsNat") {
-            return predef.ORDINAL_IsNat;
-          }
-        } else if (f.FullSanitizedName == "_System.Tuple2._0") {
-          return predef.Tuple2Destructors0;
-        } else if (f.FullSanitizedName == "_System.Tuple2._1") {
-          return predef.Tuple2Destructors1;
-        }
-
-        // Create a new function
-        // function f(Ref): ty;
-        List<Variable> formals = new List<Variable>();
-        if (f is ConstantField) {
-          formals.AddRange(MkTyParamFormals(GetTypeParams(f.EnclosingClass), false));
-        }
-        if (!f.IsStatic) {
-          var udt = UserDefinedType.FromTopLevelDecl(f.tok, f.EnclosingClass);
-          Bpl.Type receiverType = TrType(udt);
-          formals.Add(new Bpl.Formal(f.tok, new Bpl.TypedIdent(f.tok, f is ConstantField ? "this" : Bpl.TypedIdent.NoName, receiverType), true));
-        }
-        Bpl.Formal result = new Bpl.Formal(f.tok, new Bpl.TypedIdent(f.tok, Bpl.TypedIdent.NoName, TrType(f.Type)), false);
-        ff = new Bpl.Function(f.tok, f.FullSanitizedName, new List<TypeVariable>(), formals, result, null, null);
-
-        if (InsertChecksums) {
-          var dt = f.EnclosingClass as DatatypeDecl;
-          if (dt != null) {
-            InsertChecksum(dt, ff);
-          }
-          // TODO(wuestholz): Do we need to handle more cases?
-        }
-
-        // add the newly created function to the cache, so that there will only be one copy of it
-        fieldFunctions.Add(f, ff);
-
-        // declare function among Boogie top-level declarations, if needed, and treat certain fields specially
-        if (f is ConstantField) {
-          // declare the function with its initial value, if any
-          // function QQ():int { 3 }
-          var cf = (ConstantField)f;
-          if (cf.Rhs != null && RevealedInScope(cf)) {
-            var etran = new ExpressionTranslator(this, predef, NewOneHeapExpr(f.tok));
-            if (!IsOpaque(cf)) {
-              sink.AddTopLevelDeclaration(ff.CreateDefinitionAxiom(etran.TrExpr(cf.Rhs)));
-            }
-          }
-          sink.AddTopLevelDeclaration(ff);
-
-        } else if (f.EnclosingClass is ArrayClassDecl) {
-          // add non-negative-range axioms for array Length fields
-          // axiom (forall o: Ref :: 0 <= array.Length(o));
-          Bpl.BoundVariable oVar = new Bpl.BoundVariable(f.tok, new Bpl.TypedIdent(f.tok, "o", predef.RefType));
-          Bpl.IdentifierExpr o = new Bpl.IdentifierExpr(f.tok, oVar);
-          var rhs = new Bpl.NAryExpr(f.tok, new Bpl.FunctionCall(ff), new List<Bpl.Expr> { o });
-          Bpl.Expr body = Bpl.Expr.Le(Bpl.Expr.Literal(0), rhs);
-          var trigger = BplTrigger(rhs);
-          Bpl.Expr qq = new Bpl.ForallExpr(f.tok, new List<Variable> { oVar }, trigger, body);
-          sink.AddTopLevelDeclaration(new Bpl.Axiom(f.tok, qq));
-        }
-      }
-      return ff;
-    }
-
-    Bpl.Expr GetField(MemberSelectExpr fse) {
-      Contract.Requires(fse != null);
-      Contract.Requires(fse.Member != null && fse.Member is Field && ((Field)(fse.Member)).IsMutable);
-      Contract.Ensures(Contract.Result<Bpl.Expr>() != null);
-
-      return new Bpl.IdentifierExpr(fse.tok, GetField((Field)fse.Member));
     }
 
     /// <summary>
@@ -5195,325 +3663,6 @@ namespace Microsoft.Dafny {
       return oldBfs;
     }
 
-    /// <summary>
-    /// Emit to "builder" a check that calleeDecreases is less than contextDecreases.  More precisely,
-    /// the check is:
-    ///     allowance || (calleeDecreases LESS contextDecreases).
-    /// </summary>
-    void CheckCallTermination(IToken tok, List<Expression> contextDecreases, List<Expression> calleeDecreases,
-                              Bpl.Expr allowance,
-                              Expression receiverReplacement, Dictionary<IVariable, Expression> substMap,
-                              Dictionary<TypeParameter, Type> typeMap,
-                              ExpressionTranslator etranCurrent, ExpressionTranslator etranInitial, BoogieStmtListBuilder builder, bool inferredDecreases, string hint) {
-      Contract.Requires(tok != null);
-      Contract.Requires(cce.NonNullElements(contextDecreases));
-      Contract.Requires(cce.NonNullElements(calleeDecreases));
-      Contract.Requires(cce.NonNullDictionaryAndValues(substMap));
-      Contract.Requires(etranCurrent != null);
-      Contract.Requires(etranInitial != null);
-      Contract.Requires(builder != null);
-
-      // The interpretation of the given decreases-clause expression tuples is as a lexicographic tuple, extended into
-      // an infinite tuple by appending TOP elements.  The TOP element is strictly larger than any other value given
-      // by a Dafny expression.  Each Dafny types has its own ordering, and these orderings are combined into a partial
-      // order where elements from different Dafny types are incomparable.  Thus, as an optimization below, if two
-      // components from different types are compared, the answer is taken to be false.
-
-      if (Contract.Exists(calleeDecreases, e => e is WildcardExpr)) {
-        // no check needed
-        return;
-      }
-
-      int N = Math.Min(contextDecreases.Count, calleeDecreases.Count);
-      var toks = new List<IToken>();
-      var types0 = new List<Type>();
-      var types1 = new List<Type>();
-      var callee = new List<Expr>();
-      var caller = new List<Expr>();
-      if (RefinementToken.IsInherited(tok, currentModule) && contextDecreases.All(e => !RefinementToken.IsInherited(e.tok, currentModule))) {
-        // the call site is inherited but all the context decreases expressions are new
-        tok = new ForceCheckToken(tok);
-      }
-      for (int i = 0; i < N; i++) {
-        Expression e0 = Substitute(calleeDecreases[i], receiverReplacement, substMap, typeMap);
-        Expression e1 = contextDecreases[i];
-        if (!CompatibleDecreasesTypes(e0.Type, e1.Type)) {
-          N = i;
-          break;
-        }
-        toks.Add(new NestedToken(tok, e1.tok));
-        types0.Add(e0.Type.NormalizeExpand());
-        types1.Add(e1.Type.NormalizeExpand());
-        callee.Add(etranCurrent.TrExpr(e0));
-        caller.Add(etranInitial.TrExpr(e1));
-      }
-      bool endsWithWinningTopComparison = N == contextDecreases.Count && N < calleeDecreases.Count;
-      Bpl.Expr decrExpr = DecreasesCheck(toks, types0, types1, callee, caller, builder, "", endsWithWinningTopComparison, false);
-      if (allowance != null) {
-        decrExpr = Bpl.Expr.Or(allowance, decrExpr);
-      }
-      builder.Add(Assert(tok, decrExpr, new PODesc.Terminates(inferredDecreases, false, hint)));
-    }
-
-    /// <summary>
-    /// Returns the expression that says whether or not the decreases function has gone down (if !allowNoChange)
-    /// or has gone down or stayed the same (if allowNoChange).
-    /// ee0 represents the new values and ee1 represents old values.
-    /// If builder is non-null, then the check '0 ATMOST decr' is generated to builder.
-    /// Requires all types in types0 and types1 to be non-proxy non-synonym types (that is, callers should invoke NormalizeExpand)
-    /// </summary>
-    Bpl.Expr DecreasesCheck(List<IToken> toks, List<Type> types0, List<Type> types1, List<Bpl.Expr> ee0, List<Bpl.Expr> ee1,
-                            BoogieStmtListBuilder builder, string suffixMsg, bool allowNoChange, bool includeLowerBound) {
-      Contract.Requires(cce.NonNullElements(toks));
-      Contract.Requires(cce.NonNullElements(types0));
-      Contract.Requires(cce.NonNullElements(types1));
-      Contract.Requires(cce.NonNullElements(ee0));
-      Contract.Requires(cce.NonNullElements(ee1));
-      Contract.Requires(predef != null);
-      Contract.Requires(types0.Count == types1.Count && types0.Count == ee0.Count && ee0.Count == ee1.Count);
-      Contract.Requires(builder == null || suffixMsg != null);
-      Contract.Ensures(Contract.Result<Bpl.Expr>() != null);
-
-      int N = types0.Count;
-
-      // compute eq and less for each component of the lexicographic tuple
-      List<Bpl.Expr> Eq = new List<Bpl.Expr>(N);
-      List<Bpl.Expr> Less = new List<Bpl.Expr>(N);
-      for (int i = 0; i < N; i++) {
-        Bpl.Expr less, atmost, eq;
-        ComputeLessEq(toks[i], types0[i], types1[i], ee0[i], ee1[i], out less, out atmost, out eq, includeLowerBound);
-        Eq.Add(eq);
-        Less.Add(allowNoChange ? atmost : less);
-      }
-      if (builder != null) {
-        // check: 0 <= ee1
-        // more precisely, for component k of the lexicographic decreases function, check:
-        //   ee0[0] < ee1[0] || ee0[1] < ee1[1] || ... || ee0[k-1] < ee1[k-1] || ee0[k] == ee1[k] || 0 <= ee1[k]
-        for (int k = 0; k < N; k++) {
-          // we only need to check lower bound for integers--sets, sequences, booleans, references, and datatypes all have natural lower bounds
-          Bpl.Expr prefixIsLess = Bpl.Expr.False;
-          for (int i = 0; i < k; i++) {
-            prefixIsLess = Bpl.Expr.Or(prefixIsLess, Less[i]);
-          };
-
-          Bpl.Expr zero = null;
-          string zeroStr = null;
-          if (types0[k].IsNumericBased(Type.NumericPersuasion.Int)) {
-            zero = Bpl.Expr.Literal(0);
-            zeroStr = "0";
-          } else if (types0[k].IsNumericBased(Type.NumericPersuasion.Real)) {
-            zero = Bpl.Expr.Literal(BaseTypes.BigDec.ZERO);
-            zeroStr = "0.0";
-          }
-          if (zero != null) {
-            Bpl.Expr bounded = Bpl.Expr.Le(zero, ee1[k]);
-            for (int i = 0; i < k; i++) {
-              bounded = Bpl.Expr.Or(bounded, Less[i]);
-            }
-            Bpl.Cmd cmd = Assert(toks[k], Bpl.Expr.Or(bounded, Eq[k]), new PODesc.DecreasesBoundedBelow(N, k, zeroStr, suffixMsg));
-            builder.Add(cmd);
-          }
-        }
-      }
-      // check: ee0 < ee1 (or ee0 <= ee1, if allowNoChange)
-      Bpl.Expr decrCheck = allowNoChange ? Bpl.Expr.True : Bpl.Expr.False;
-      for (int i = N; 0 <= --i;) {
-        Bpl.Expr less = Less[i];
-        Bpl.Expr eq = Eq[i];
-        if (allowNoChange) {
-          // decrCheck = atmost && (eq ==> decrCheck)
-          decrCheck = Bpl.Expr.And(less, Bpl.Expr.Imp(eq, decrCheck));
-        } else {
-          // decrCheck = less || (eq && decrCheck)
-          decrCheck = Bpl.Expr.Or(less, Bpl.Expr.And(eq, decrCheck));
-        }
-      }
-      return decrCheck;
-    }
-
-    bool CompatibleDecreasesTypes(Type t, Type u) {
-      Contract.Requires(t != null);
-      Contract.Requires(u != null);
-      t = t.NormalizeExpand();
-      u = u.NormalizeExpand();
-      if (t is BoolType) {
-        return u is BoolType;
-      } else if (t is CharType) {
-        return u is CharType;
-      } else if (t.IsNumericBased(Type.NumericPersuasion.Int)) {
-        // we can allow different kinds of int-based types
-        return u.IsNumericBased(Type.NumericPersuasion.Int);
-      } else if (t.IsNumericBased(Type.NumericPersuasion.Real)) {
-        // we can allow different kinds of real-based types
-        return u.IsNumericBased(Type.NumericPersuasion.Real);
-      } else if (t is SetType) {
-        return u is SetType;
-      } else if (t is SeqType) {
-        return u is SeqType || u.IsIndDatatype;
-      } else if (t.IsDatatype) {
-        return u.IsDatatype || (t.IsIndDatatype && u is SeqType);
-      } else if (t.IsRefType) {
-        return u.IsRefType;
-      } else if (t is MultiSetType) {
-        return u is MultiSetType;
-      } else if (t is MapType) {
-        return u is MapType && ((MapType)t).Finite == ((MapType)u).Finite;
-      } else if (t is ArrowType) {
-        return u is ArrowType;
-      } else if (t is BitvectorType) {
-        return u is BitvectorType;
-      } else if (t is BigOrdinalType) {
-        return u is BigOrdinalType;
-      } else {
-        Contract.Assert(t.IsTypeParameter || t.IsAbstractType || t.IsInternalTypeSynonym);
-        return false;  // don't consider any type parameters to be the same (since we have no comparison function for them anyway)
-      }
-    }
-
-    Nullable<BuiltinFunction> RankFunction(Type/*!*/ ty) {
-      Contract.Requires(ty != null);
-      if (ty is SeqType) {
-        return BuiltinFunction.SeqRank;
-      } else if (ty.IsIndDatatype) {
-        return BuiltinFunction.DtRank;
-      } else {
-        return null;
-      }
-    }
-
-    void ComputeLessEq(IToken tok, Type ty0, Type ty1, Bpl.Expr e0, Bpl.Expr e1, out Bpl.Expr less, out Bpl.Expr atmost, out Bpl.Expr eq, bool includeLowerBound) {
-      Contract.Requires(tok != null);
-      Contract.Requires(ty0 != null);
-      Contract.Requires(ty1 != null);
-      Contract.Requires(e0 != null);
-      Contract.Requires(e1 != null);
-      Contract.Requires(predef != null);
-      Contract.Ensures(Contract.ValueAtReturn(out less) != null);
-      Contract.Ensures(Contract.ValueAtReturn(out atmost) != null);
-      Contract.Ensures(Contract.ValueAtReturn(out eq) != null);
-
-      ty0 = ty0.NormalizeExpand();
-      ty1 = ty1.NormalizeExpand();
-      var rk0 = RankFunction(ty0);
-      var rk1 = RankFunction(ty1);
-      if (rk0 != null && rk1 != null && rk0 != rk1) {
-        eq = Bpl.Expr.False;
-        Bpl.Expr b0 = FunctionCall(tok, rk0.Value, null, e0);
-        Bpl.Expr b1 = FunctionCall(tok, rk1.Value, null, e1);
-        less = Bpl.Expr.Lt(b0, b1);
-        atmost = Bpl.Expr.Le(b0, b1);
-      } else if (ty0 is BoolType) {
-        eq = Bpl.Expr.Iff(e0, e1);
-        less = Bpl.Expr.And(Bpl.Expr.Not(e0), e1);
-        atmost = Bpl.Expr.Imp(e0, e1);
-      } else if (ty0 is CharType) {
-        eq = Bpl.Expr.Eq(e0, e1);
-        var operand0 = FunctionCall(e0.tok, BuiltinFunction.CharToInt, null, e0);
-        var operand1 = FunctionCall(e0.tok, BuiltinFunction.CharToInt, null, e1);
-        less = Bpl.Expr.Binary(tok, BinaryOperator.Opcode.Lt, operand0, operand1);
-        atmost = Bpl.Expr.Binary(tok, BinaryOperator.Opcode.Le, operand0, operand1);
-      } else if (ty0.IsNumericBased(Type.NumericPersuasion.Int) || ty0 is SeqType || ty0.IsDatatype) {
-        Bpl.Expr b0, b1;
-        if (ty0.IsNumericBased(Type.NumericPersuasion.Int)) {
-          b0 = e0;
-          b1 = e1;
-        } else if (ty0 is SeqType) {
-          b0 = FunctionCall(tok, BuiltinFunction.SeqRank, null, e0);
-          b1 = FunctionCall(tok, BuiltinFunction.SeqRank, null, e1);
-        } else if (ty0.IsDatatype) {
-          b0 = FunctionCall(tok, BuiltinFunction.DtRank, null, e0);
-          b1 = FunctionCall(tok, BuiltinFunction.DtRank, null, e1);
-        } else {
-          Contract.Assert(false); throw new cce.UnreachableException();
-        }
-        eq = Bpl.Expr.Eq(b0, b1);
-        less = Bpl.Expr.Lt(b0, b1);
-        atmost = Bpl.Expr.Le(b0, b1);
-        if (ty0.IsNumericBased(Type.NumericPersuasion.Int) && includeLowerBound) {
-          less = Bpl.Expr.And(Bpl.Expr.Le(Bpl.Expr.Literal(0), b0), less);
-          atmost = Bpl.Expr.And(Bpl.Expr.Le(Bpl.Expr.Literal(0), b0), atmost);
-        }
-
-      } else if (ty0.IsNumericBased(Type.NumericPersuasion.Real)) {
-        eq = Bpl.Expr.Eq(e0, e1);
-        less = Bpl.Expr.Le(e0, Bpl.Expr.Sub(e1, Bpl.Expr.Literal(BaseTypes.BigDec.FromInt(1))));
-        atmost = Bpl.Expr.Le(e0, e1);
-        if (includeLowerBound) {
-          less = Bpl.Expr.And(Bpl.Expr.Le(Bpl.Expr.Literal(BaseTypes.BigDec.ZERO), e0), less);
-          atmost = Bpl.Expr.And(Bpl.Expr.Le(Bpl.Expr.Literal(BaseTypes.BigDec.ZERO), e0), atmost);
-        }
-
-      } else if (ty0 is IteratorDecl.EverIncreasingType) {
-        eq = Bpl.Expr.Eq(e0, e1);
-        less = Bpl.Expr.Gt(e0, e1);
-        atmost = Bpl.Expr.Ge(e0, e1);
-
-      } else if ((ty0 is SetType && ((SetType)ty0).Finite) || (ty0 is MapType && ((MapType)ty0).Finite)) {
-        Bpl.Expr b0, b1;
-        if (ty0 is SetType) {
-          b0 = e0;
-          b1 = e1;
-        } else {
-          // for maps, compare their domains as sets
-          b0 = FunctionCall(tok, BuiltinFunction.MapDomain, predef.MapType(tok, true, predef.BoxType, predef.BoxType), e0);
-          b1 = FunctionCall(tok, BuiltinFunction.MapDomain, predef.MapType(tok, true, predef.BoxType, predef.BoxType), e1);
-        }
-        eq = FunctionCall(tok, BuiltinFunction.SetEqual, null, b0, b1);
-        less = ProperSubset(tok, b0, b1);
-        atmost = FunctionCall(tok, BuiltinFunction.SetSubset, null, b0, b1);
-
-      } else if (ty0 is SetType || ty0 is MapType) {
-        Bpl.Expr b0, b1;
-        if (ty0 is SetType) {
-          Contract.Assert(!((SetType)ty0).Finite);
-          b0 = e0;
-          b1 = e1;
-        } else {
-          Contract.Assert(!((MapType)ty0).Finite);
-          // for maps, compare their domains as sets
-          b0 = FunctionCall(tok, BuiltinFunction.IMapDomain, predef.MapType(tok, false, predef.BoxType, predef.BoxType), e0);
-          b1 = FunctionCall(tok, BuiltinFunction.IMapDomain, predef.MapType(tok, false, predef.BoxType, predef.BoxType), e1);
-        }
-        eq = FunctionCall(tok, BuiltinFunction.ISetEqual, null, b0, b1);
-        less = Bpl.Expr.False;
-        atmost = BplOr(less, eq);
-
-      } else if (ty0 is MultiSetType) {
-        eq = FunctionCall(tok, BuiltinFunction.MultiSetEqual, null, e0, e1);
-        less = ProperMultiset(tok, e0, e1);
-        atmost = FunctionCall(tok, BuiltinFunction.MultiSetSubset, null, e0, e1);
-
-      } else if (ty0 is ArrowType) {
-        eq = Bpl.Expr.Eq(e0, e1);
-        less = Bpl.Expr.False;  // TODO: try to do better than this
-        atmost = BplOr(less, eq);
-
-      } else if (ty0 is BitvectorType) {
-        BitvectorType bv = (BitvectorType)ty0;
-        eq = Bpl.Expr.Eq(e0, e1);
-        less = FunctionCall(tok, "lt_bv" + bv.Width, Bpl.Type.Bool, e0, e1);
-        atmost = FunctionCall(tok, "le_bv" + bv.Width, Bpl.Type.Bool, e0, e1);
-
-      } else if (ty0 is BigOrdinalType) {
-        eq = Bpl.Expr.Eq(e0, e1);
-        less = FunctionCall(tok, "ORD#Less", Bpl.Type.Bool, e0, e1);
-        atmost = BplOr(eq, less);
-
-      } else if (ty0.IsTypeParameter || ty0.IsAbstractType) {
-        eq = Bpl.Expr.Eq(e0, e1);
-        less = Bpl.Expr.False;
-        atmost = BplOr(less, eq);
-
-      } else {
-        // reference type
-        Contract.Assert(ty0.IsRefType);  // otherwise, unexpected type
-        var b0 = Bpl.Expr.Neq(e0, predef.Null);
-        var b1 = Bpl.Expr.Neq(e1, predef.Null);
-        eq = Bpl.Expr.Iff(b0, b1);
-        less = Bpl.Expr.And(Bpl.Expr.Not(b0), b1);
-        atmost = Bpl.Expr.Imp(b0, b1);
-      }
-    }
 
     void AddComment(BoogieStmtListBuilder builder, Statement stmt, string comment) {
       Contract.Requires(builder != null);
@@ -5911,328 +4060,6 @@ namespace Microsoft.Dafny {
     }
 
     int projectionFunctionCount = 0;
-
-    /// <summary>
-    /// Fills in, if necessary, the e.translationDesugaring field, and returns it.
-    /// Also, makes sure that letSuchThatExprInfo maps e to something.
-    /// </summary>
-    Expression LetDesugaring(LetExpr e) {
-      Contract.Requires(e != null);
-      Contract.Requires(!e.Exact);
-      Contract.Ensures(Contract.Result<Expression>() != null);
-      if (e.GetTranslationDesugaring(this) == null) {
-        // For let-such-that expression:
-        //   var x:X, y:Y :| P(x,y,g); F(...)
-        // where
-        //   - g has type G, and
-        //   - tt* denotes the list of type variables in the types X and Y and expression F(...),
-        // declare a function for each bound variable:
-        //   function $let$x(Ty*, G): X;
-        //   function $let$y(Ty*, G): Y;
-        //   function $let_canCall(Ty*, G): bool;
-        // and add an axiom about these functions:
-        //   axiom (forall tt*:Ty*, g:G ::
-        //            { $let$x(tt*, g) }
-        //            { $let$y(tt*, g) }
-        //            $let$_canCall(tt*, g)) ==>
-        //            P($let$x(tt*, g), $let$y(tt*, g), g));
-        // and create the desugaring:
-        //   var x:X, y:Y := $let$x(tt*, g), $let$y(tt*, g); F(...)
-        if (e is SubstLetExpr) {
-          // desugar based on the original letexpr.
-          var expr = (SubstLetExpr)e;
-          var orgExpr = expr.orgExpr;
-          Expression d = LetDesugaring(orgExpr);
-          e.SetTranslationDesugaring(this, Substitute(d, null, expr.substMap, expr.typeMap));
-          var orgInfo = letSuchThatExprInfo[orgExpr];
-          letSuchThatExprInfo.Add(expr, new LetSuchThatExprInfo(orgInfo, this, expr.substMap, expr.typeMap));
-        } else {
-          // First, determine "g" as a list of Dafny variables FVs plus possibly this, $Heap, and old($Heap),
-          // and determine "tt*" as a list of Dafny type variables
-          LetSuchThatExprInfo info;
-          {
-            var FVs = new HashSet<IVariable>();
-            bool usesHeap = false, usesOldHeap = false;
-            var FVsHeapAt = new HashSet<Label>();
-            Type usesThis = null;
-            FreeVariablesUtil.ComputeFreeVariables(options, e.RHSs[0], FVs, ref usesHeap, ref usesOldHeap, FVsHeapAt, ref usesThis, false);
-            var FTVs = new HashSet<TypeParameter>();
-            foreach (var bv in e.BoundVars) {
-              FVs.Remove(bv);
-              ComputeFreeTypeVariables(bv.Type, FTVs);
-            }
-            ComputeFreeTypeVariables(e.RHSs[0], FTVs);
-            info = new LetSuchThatExprInfo(e.tok, letSuchThatExprInfo.Count, FVs.ToList(), FTVs.ToList(), usesHeap, usesOldHeap, FVsHeapAt, usesThis, currentDeclaration);
-            letSuchThatExprInfo.Add(e, info);
-          }
-
-          foreach (var bv in e.BoundVars) {
-            Bpl.Variable resType = new Bpl.Formal(bv.tok, new Bpl.TypedIdent(bv.tok, Bpl.TypedIdent.NoName, TrType(bv.Type)), false);
-            Bpl.Expr ante;
-            List<Variable> formals = info.GAsVars(this, true, out ante, null);
-            var fn = new Bpl.Function(bv.tok, info.SkolemFunctionName(bv), formals, resType);
-
-            if (InsertChecksums) {
-              InsertChecksum(e.Body, fn);
-            }
-
-            sink.AddTopLevelDeclaration(fn);
-          }
-          var canCallFunction = AddLetSuchThatCanCallFunction(e, info);
-          AddLetSuchThenCanCallAxiom(e, info, canCallFunction);
-
-          // now that we've declared the functions and axioms, let's prepare the let-such-that desugaring
-          {
-            var etran = new ExpressionTranslator(this, predef, e.tok);
-            var rhss = new List<Expression>();
-            foreach (var bv in e.BoundVars) {
-              var args = info.SkolemFunctionArgs(bv, this, etran);
-              var rhs = new BoogieFunctionCall(bv.tok, info.SkolemFunctionName(bv), info.UsesHeap, info.UsesOldHeap, info.UsesHeapAt, args.Item1, args.Item2);
-              rhs.Type = bv.Type;
-              rhss.Add(rhs);
-            }
-            var expr = new LetExpr(e.tok, e.LHSs, rhss, e.Body, true);
-            expr.Type = e.Type; // resolve here
-            e.SetTranslationDesugaring(this, expr);
-          }
-        }
-      }
-      return e.GetTranslationDesugaring(this);
-    }
-
-    private Bpl.Function AddLetSuchThatCanCallFunction(LetExpr e, LetSuchThatExprInfo info) {
-      Bpl.Variable resType = new Bpl.Formal(e.tok, new Bpl.TypedIdent(e.tok, Bpl.TypedIdent.NoName, Bpl.Type.Bool),
-        false);
-      List<Variable> formals = info.GAsVars(this, true, out var ante, null);
-      var canCallFunction = new Bpl.Function(e.tok, info.CanCallFunctionName(), formals, resType);
-
-      if (InsertChecksums) {
-        InsertChecksum(e.Body, canCallFunction);
-      }
-
-      sink.AddTopLevelDeclaration(canCallFunction);
-      return canCallFunction;
-    }
-
-    private void AddLetSuchThenCanCallAxiom(LetExpr e, LetSuchThatExprInfo info, Bpl.Function canCallFunction) {
-      var etranCC = new ExpressionTranslator(this, predef, info.HeapExpr(this, false), info.HeapExpr(this, true));
-      Bpl.Expr typeAntecedents; // later ignored
-      List<Variable> gg = info.GAsVars(this, false, out typeAntecedents, etranCC);
-      var gExprs = new List<Bpl.Expr>();
-      foreach (Bpl.Variable g in gg) {
-        gExprs.Add(new Bpl.IdentifierExpr(g.tok, g));
-      }
-
-      Bpl.Trigger tr = null;
-      Dictionary<IVariable, Expression> substMap = new Dictionary<IVariable, Expression>();
-      Bpl.Expr antecedent = Bpl.Expr.True;
-      foreach (var bv in e.BoundVars) {
-        // create a call to $let$x(g)
-        var call = FunctionCall(e.tok, info.SkolemFunctionName(bv), TrType(bv.Type), gExprs);
-        tr = new Bpl.Trigger(e.tok, true, new List<Bpl.Expr> { call }, tr);
-        substMap.Add(bv, new BoogieWrapper(call, bv.Type));
-        if (!(bv.Type.IsTypeParameter)) {
-          Bpl.Expr wh = GetWhereClause(bv.tok, call, bv.Type, etranCC, NOALLOC);
-          if (wh != null) {
-            antecedent = BplAnd(antecedent, wh);
-          }
-        }
-      }
-
-      var i = info.FTVs.Count + (info.UsesHeap ? 1 : 0) + (info.UsesOldHeap ? 1 : 0) + info.UsesHeapAt.Count;
-      Expression receiverReplacement;
-      if (info.ThisType == null) {
-        receiverReplacement = null;
-      } else {
-        receiverReplacement = new BoogieWrapper(gExprs[i], info.ThisType);
-        i++;
-      }
-
-      foreach (var fv in info.FVs) {
-        var ge = gExprs[i];
-        substMap.Add(fv, new BoogieWrapper(ge, fv.Type));
-        i++;
-      }
-
-      var canCall = FunctionCall(e.tok, info.CanCallFunctionName(), Bpl.Type.Bool, gExprs);
-      var p = Substitute(e.RHSs[0], receiverReplacement, substMap);
-      Bpl.Expr ax = Bpl.Expr.Imp(canCall, BplAnd(antecedent, etranCC.TrExpr(p)));
-      ax = BplForall(gg, tr, ax);
-      AddOtherDefinition(canCallFunction, new Bpl.Axiom(e.tok, ax));
-    }
-
-    class LetSuchThatExprInfo {
-      public readonly IToken Tok;
-      public readonly int LetId;
-      public readonly List<IVariable> FVs;
-      public readonly List<Expression> FV_Exprs;  // these are what initially were the free variables, but they may have undergone substitution so they are here Expression's.
-      public readonly List<TypeParameter> FTVs;
-      public readonly List<Type> FTV_Types;
-      public readonly bool UsesHeap;
-      public readonly bool UsesOldHeap;
-      public readonly List<Label> UsesHeapAt;
-      public readonly Type ThisType;  // null if 'this' is not used
-      public LetSuchThatExprInfo(IToken tok, int uniqueLetId,
-      List<IVariable> freeVariables, List<TypeParameter> freeTypeVars,
-      bool usesHeap, bool usesOldHeap, ISet<Label> usesHeapAt, Type thisType, Declaration currentDeclaration) {
-        Tok = tok;
-        LetId = uniqueLetId;
-        FTVs = freeTypeVars;
-        FTV_Types = Map(freeTypeVars, tt => (Type)new UserDefinedType(tt));
-        FVs = freeVariables;
-        FV_Exprs = new List<Expression>();
-        foreach (var v in FVs) {
-          var idExpr = new IdentifierExpr(v.Tok, v.AssignUniqueName(currentDeclaration.IdGenerator));
-          idExpr.Var = v; idExpr.Type = v.Type;  // resolve here
-          FV_Exprs.Add(idExpr);
-        }
-        UsesHeap = usesHeap;
-        UsesOldHeap = usesOldHeap;
-        // we convert the set of heap-at variables to a list here, once and for all; the order itself is not material, what matters is that we always use the same order
-        UsesHeapAt = new List<Label>(usesHeapAt);
-        ThisType = thisType;
-      }
-      public LetSuchThatExprInfo(LetSuchThatExprInfo template, BoogieGenerator boogieGenerator,
-           Dictionary<IVariable, Expression> substMap,
-           Dictionary<TypeParameter, Type> typeMap) {
-        Contract.Requires(template != null);
-        Contract.Requires(boogieGenerator != null);
-        Contract.Requires(substMap != null);
-        Tok = template.Tok;
-        LetId = template.LetId;  // reuse the ID, which ensures we get the same $let functions
-        FTVs = template.FTVs;
-        FTV_Types = template.FTV_Types.ConvertAll(t => t.Subst(typeMap));
-        FVs = template.FVs;
-        FV_Exprs = template.FV_Exprs.ConvertAll(e => BoogieGenerator.Substitute(e, null, substMap, typeMap));
-        UsesHeap = template.UsesHeap;
-        UsesOldHeap = template.UsesOldHeap;
-        UsesHeapAt = template.UsesHeapAt;
-        ThisType = template.ThisType;
-      }
-      public Tuple<List<Expression>, List<Type>> SkolemFunctionArgs(BoundVar bv, BoogieGenerator boogieGenerator, ExpressionTranslator etran) {
-        Contract.Requires(bv != null);
-        Contract.Requires(boogieGenerator != null);
-        Contract.Requires(etran != null);
-        var args = new List<Expression>();
-        if (ThisType != null) {
-          var th = new ThisExpr(bv.tok);
-          th.Type = ThisType;
-          args.Add(th);
-        }
-        args.AddRange(FV_Exprs);
-        return Tuple.Create(args, new List<Type>(FTV_Types));
-      }
-      public string SkolemFunctionName(BoundVar bv) {
-        Contract.Requires(bv != null);
-        return string.Format("$let#{0}_{1}", LetId, bv.Name);
-      }
-      public Bpl.Expr CanCallFunctionCall(BoogieGenerator boogieGenerator, ExpressionTranslator etran) {
-        Contract.Requires(boogieGenerator != null);
-        Contract.Requires(etran != null);
-        var gExprs = new List<Bpl.Expr>();
-        gExprs.AddRange(Map(FTV_Types, tt => boogieGenerator.TypeToTy(tt)));
-        if (UsesHeap) {
-          gExprs.Add(etran.HeapExpr);
-        }
-        if (UsesOldHeap) {
-          gExprs.Add(etran.Old.HeapExpr);
-        }
-        foreach (var heapAtLabel in UsesHeapAt) {
-          Bpl.Expr ve;
-          var bv = BplBoundVar("$Heap_at_" + heapAtLabel.AssignUniqueId(boogieGenerator.CurrentIdGenerator), boogieGenerator.predef.HeapType, out ve);
-          gExprs.Add(ve);
-        }
-        if (ThisType != null) {
-          var th = new Bpl.IdentifierExpr(Tok, etran.This);
-          gExprs.Add(th);
-        }
-        foreach (var v in FV_Exprs) {
-          gExprs.Add(etran.TrExpr(v));
-        }
-        return FunctionCall(Tok, CanCallFunctionName(), Bpl.Type.Bool, gExprs);
-      }
-      public string CanCallFunctionName() {
-        return string.Format("$let#{0}$canCall", LetId);
-      }
-      public Bpl.Expr HeapExpr(BoogieGenerator boogieGenerator, bool old) {
-        Contract.Requires(boogieGenerator != null);
-        return new Bpl.IdentifierExpr(Tok, old ? "$heap$old" : "$heap", boogieGenerator.predef.HeapType);
-      }
-      /// <summary>
-      /// "wantFormals" means the returned list will consist of all in-parameters.
-      /// "!wantFormals" means the returned list will consist of all bound variables.
-      /// Guarantees that, in the list returned, "this" is the parameter immediately following
-      /// the (0, 1, or 2) heap arguments, if there is a "this" parameter at all.
-      /// Note, "typeAntecedents" is meaningfully filled only if "etran" is not null.
-      /// </summary>
-      public List<Variable> GAsVars(BoogieGenerator boogieGenerator, bool wantFormals, out Bpl.Expr typeAntecedents, ExpressionTranslator etran) {
-        Contract.Requires(boogieGenerator != null);
-        var vv = new List<Variable>();
-        // first, add the type variables
-        vv.AddRange(Map(FTVs, tp => NewVar(NameTypeParam(tp), boogieGenerator.predef.Ty, wantFormals)));
-        typeAntecedents = Bpl.Expr.True;
-        if (UsesHeap) {
-          var nv = NewVar("$heap", boogieGenerator.predef.HeapType, wantFormals);
-          vv.Add(nv);
-          if (etran != null) {
-            var isGoodHeap = boogieGenerator.FunctionCall(Tok, BuiltinFunction.IsGoodHeap, null, new Bpl.IdentifierExpr(Tok, nv));
-            typeAntecedents = BplAnd(typeAntecedents, isGoodHeap);
-          }
-        }
-        if (UsesOldHeap) {
-          var nv = NewVar("$heap$old", boogieGenerator.predef.HeapType, wantFormals);
-          vv.Add(nv);
-          if (etran != null) {
-            var isGoodHeap = boogieGenerator.FunctionCall(Tok, BuiltinFunction.IsGoodHeap, null, new Bpl.IdentifierExpr(Tok, nv));
-            typeAntecedents = BplAnd(typeAntecedents, isGoodHeap);
-          }
-        }
-        foreach (var heapAtLabel in UsesHeapAt) {
-          var nv = NewVar("$Heap_at_" + heapAtLabel.AssignUniqueId(boogieGenerator.CurrentIdGenerator), boogieGenerator.predef.HeapType, wantFormals);
-          vv.Add(nv);
-          if (etran != null) {
-            // TODO: It's not clear to me that $IsGoodHeap predicates are needed for these axioms. (Same comment applies above for $heap$old.)
-            // But $HeapSucc relations among the various heap variables appears not needed for either soundness or completeness, since the
-            // let-such-that functions will always be invoked on arguments for which these properties are known.
-            var isGoodHeap = boogieGenerator.FunctionCall(Tok, BuiltinFunction.IsGoodHeap, null, new Bpl.IdentifierExpr(Tok, nv));
-            typeAntecedents = BplAnd(typeAntecedents, isGoodHeap);
-          }
-        }
-        if (ThisType != null) {
-          var nv = NewVar("this", boogieGenerator.TrType(ThisType), wantFormals);
-          vv.Add(nv);
-          if (etran != null) {
-            var th = new Bpl.IdentifierExpr(Tok, nv);
-            typeAntecedents = BplAnd(typeAntecedents, boogieGenerator.ReceiverNotNull(th));
-            var wh = boogieGenerator.GetWhereClause(Tok, th, ThisType, etran, NOALLOC);
-            if (wh != null) {
-              typeAntecedents = BplAnd(typeAntecedents, wh);
-            }
-          }
-        }
-        foreach (var v in FVs) {
-          var nv = NewVar(v.Name, boogieGenerator.TrType(v.Type), wantFormals);
-          vv.Add(nv);
-          if (etran != null) {
-            var wh = boogieGenerator.GetWhereClause(Tok, new Bpl.IdentifierExpr(Tok, nv), v.Type, etran, NOALLOC);
-            if (wh != null) {
-              typeAntecedents = BplAnd(typeAntecedents, wh);
-            }
-          }
-        }
-        return vv;
-      }
-      Bpl.Variable NewVar(string name, Bpl.Type type, bool wantFormal) {
-        Contract.Requires(name != null);
-        Contract.Requires(type != null);
-        if (wantFormal) {
-          return new Bpl.Formal(Tok, new Bpl.TypedIdent(Tok, name, type), true);
-        } else {
-          return new Bpl.BoundVariable(Tok, new Bpl.TypedIdent(Tok, name, type));
-        }
-      }
-    }
-    Dictionary<LetExpr, LetSuchThatExprInfo> letSuchThatExprInfo = new Dictionary<LetExpr, LetSuchThatExprInfo>();
     private Declaration currentDeclaration;
 
     // ----- Expression ---------------------------------------------------------------------------
